### 2020-10-02 #2718 @whedon
[PRE REVIEW]: PyQMRI: An accelerated Python based Quantitative MRI Python toolbox
**Submitting author:** @maieroli2010 (<a href="http://orcid.org/0000-0002-7800-0022">Oliver Maier</a>)
**Repository:** <a href="https://github.com/IMTtugraz/PyQMRI" target ="_blank">https://github.com/IMTtugraz/PyQMRI</a>
**Version:** v0.3.1
