### 2020-10-02 #2718 @Kevin-Mattheus-Moerman
@openjournals/dev can you help troubleshoot why this paper is not compiling? Thanks
