### 2020-10-02 #2718 @Kevin-Mattheus-Moerman
> @openjournals/dev can you help troubleshoot why this paper is not compiling? Thanks

I think I fixed it by changing the repository link to be the repo itself not the branch
