### 2020-10-03 #2718 @MaierOli2010
I would suggest hanayik as potential reviewer given his knowledge in MRI and Image Analysis and sbacchio for his knowledge in HPC and Iterative Linear Solvers. vitorsr, ulvgard, and stsievert might be suitable too given their background in optimization and image processing.
