### 2020-10-03 #2718 @Kevin-Mattheus-Moerman
@grlee77, @JoshKarpel, @rljacobson, @mfroeling, @62442katieb, @agahkarakuzu, @mathieuboudreau, @spinicist, @imr-framework, @kerkelae , @DARSakthi, @ritagnunes

Would you be interested in reviewing this work for JOSS?

_PyQMRI: An accelerated Python based Quantitative MRI Python toolbox_
**paper** :point_right: https://github.com/openjournals/joss-papers/blob/joss.02718/joss.02718/10.21105.joss.02718.pdf
**software** :point_right: https://github.com/IMTtugraz/PyQMRI
