### 2020-10-05 53b5260
Fixed typos in JOSS manuskript and replaced time.clock() with time.perf_counter() in _helper_fun/_nlinvns.py. Added a section about Contributing to README.rst and gave an example how to run the tests.
