### 2020-10-05 #2718 @DARSakthi
I can try to arrange for GPU access. If another reviewer is needed

In general, and as an item of interest for the authors, GPUs are hard to
come by -- including solutions like Google Collab - based scripts, which
allows for free GPU access, could be valuable in expanding the user base.

Let me know if no one else steps up @Kevin-Mattheus-Moerman. Looks like an
interesting paper

Cheers

On Mon, 5 Oct 2020 at 15:31, Mathieu Boudreau <notifications@github.com>
wrote:

> Since @agahkarakuzu <https://github.com/agahkarakuzu> and I are from the
> same lab, I yield to him and decline for myself.
>
