### 2020-10-07 #2727 @Kevin-Mattheus-Moerman
@openjournals/dev @arfon have a look at this issue :point_up:, it looks like when I assigned agahkarakuzu as reviewer over at https://github.com/openjournals/joss-reviews/issues/2718 they got a double `@` in their assigned name, see also at the top of this issue. Looks like that might be causing this.
How will I fix this? Will I remove `@ @agahkarakuzu`, and add `@agahkarakuzu` as reviewer again?
