### 2020-10-07 e13b3fa
Update paper.md
Removed 2nd Python in title
