### 2020-10-08 675b3c0
Updated testcases to allow for arbitrary root directory starts.
