+++ pyqmri/kernels/OpenCL_Kernels_double_streamed.c
@@ -1,705 +1,1017 @@
-
-__kernel void update_v(__global double8 *v,__global double8 *v_, __global double8 *Kyk2, const double tau) {
-  size_t i = get_global_id(0);
-  v[i] = v_[i]-tau*Kyk2[i];
+__kernel void update_v(
+                __global double8 *v,
+                __global double8 *v_,
+                __global double8 *Kyk2,
+                const double tau
+                )
+{
+    size_t i = get_global_id(0);
+    v[i] = v_[i]-tau*Kyk2[i];
 }
 
-__kernel void update_r(__global double2 *r, __global double2 *r_, __global double2 *A, __global double2 *A_, __global double2 *res,
-                          const double sigma, const double theta, const double lambdainv) {
-  size_t i = get_global_id(0);
 
-  r[i] = (r_[i]+sigma*((1+theta)*A[i]-theta*A_[i] - res[i]))*lambdainv;
+__kernel void update_r(
+                __global double2 *r,
+                __global double2 *r_,
+                __global double2 *A,
+                __global double2 *A_,
+                __global double2 *res,
+                const double sigma,
+                const double theta,
+                const double lambdainv
+                )
+{
+    size_t i = get_global_id(0);
+    r[i] = (r_[i]+sigma*((1+theta)*A[i]-theta*A_[i] - res[i]))*lambdainv;
 }
-__kernel void update_z2(__global double16 *z_new, __global double16 *z, __global double16 *gx,__global double16 *gx_,
-                          const double sigma, const double theta, const double alphainv,
-                          const int NUk) {
-  size_t Nx = get_global_size(2), Ny = get_global_size(1);
-  size_t NSl = get_global_size(0);
-  size_t x = get_global_id(2), y = get_global_id(1);
-  size_t k = get_global_id(0);
-  size_t i = k*Nx*Ny*NUk+Nx*y + x;
 
-  double fac = 0.0f;
 
-  for (int uk=0; uk<NUk; uk++)
-  {
-     z_new[i] = z[i] + sigma*((1+theta)*gx[i]-theta*gx_[i]);
+__kernel void update_z2(
+                __global double16 *z_new,
+                __global double16 *z,
+                __global double16 *gx,
+                __global double16 *gx_,
+                const double sigma,
+                const double theta,
+                const double alphainv,
+                const int NUk
+                )
+{
+    size_t Nx = get_global_size(2), Ny = get_global_size(1);
+    size_t NSl = get_global_size(0);
+    size_t x = get_global_id(2), y = get_global_id(1);
+    size_t k = get_global_id(0);
+    size_t i = k*Nx*Ny*NUk+Nx*y + x;
 
-     // reproject
-     fac = hypot(fac,hypot(
-     hypot(hypot(hypot(z_new[i].s0,z_new[i].s1), hypot(z_new[i].s2,z_new[i].s3)),hypot(z_new[i].s4,z_new[i].s5)),
-     hypot(hypot(2.0f*hypot(z_new[i].s6,z_new[i].s7),2.0f*hypot(z_new[i].s8,z_new[i].s9)),2.0f*hypot(z_new[i].sa,z_new[i].sb)))*alphainv);
-     i+=Nx*Ny;
-   }
-  i = k*Nx*Ny*NUk+Nx*y + x;
-  for (int uk=0; uk<NUk; uk++)
-  {
-    if (fac > 1.0f) {z_new[i] /=fac;}
+    double fac = 0.0f;
+
+    for (int uk=0; uk<NUk; uk++)
+    {
[…]
