+++ pyqmri/kernels/OpenCL_Kernels_streamed.c
@@ -34,63 +34,70 @@ __kernel void update_z2(
                 const float sigma,
                 const float theta,
                 const float alphainv,
-                const int NUk
+                const int NUk,
+                const int dimZ,
+                const int dimY,
+                const int dimX
                 )
 {
-    size_t Nx = get_global_size(2), Ny = get_global_size(1);
-    size_t NSl = get_global_size(0);
-    size_t x = get_global_id(2), y = get_global_id(1);
-    size_t k = get_global_id(0);
-    size_t i = k*Nx*Ny*NUk+Nx*y + x;
-
+    size_t i = get_global_id(0);
+    size_t size = get_global_size(0);
+    
+    size_t ind = 0;
+    size_t ind_slice = i/(dimX*dimY);
+    size_t ind_xy = i%(dimX*dimY);
+    
     float fac = 0.0f;
 
     for (int uk=0; uk<NUk; uk++)
     {
-       z_new[i] = z[i] + sigma*((1+theta)*gx[i]-theta*gx_[i]);
+        ind = ind_slice*(dimX*dimY*NUk) + uk*dimX*dimY + ind_xy;
+        z_new[ind] = z[ind] + sigma*((1+theta)*gx[ind]-theta*gx_[ind]);
 
-       // reproject
-       fac = hypot(fac,hypot(
-           hypot(
-             hypot(
-               hypot(
-                 z_new[i].s0,
-                 z_new[i].s1
-                 ),
-             hypot(
-               z_new[i].s2,
-               z_new[i].s3
-               )
-             ),
-           hypot(
-             z_new[i].s4,
-             z_new[i].s5
-             )
-           ),
-         hypot(
-           hypot(
-             2.0f*hypot(
-               z_new[i].s6,
-               z_new[i].s7
-               ),
-             2.0f*hypot(
-               z_new[i].s8,
-               z_new[i].s9
-               )
-             ),
-           2.0f*hypot(
-             z_new[i].sa,
-             z_new[i].sb
-             )
-           )
-         )*alphainv);
-       i+=Nx*Ny;
+        // reproject
+        fac = hypot(fac,
+        hypot(
+          hypot(
+            hypot(
+              hypot(
+                z_new[ind].s0,
+                z_new[ind].s1
+                ),
+                hypot(
+                  z_new[ind].s2,
+                  z_new[ind].s3
+                  )
+              ),
+              hypot(
+                z_new[ind].s4
+                ,z_new[ind].s5
+                )
+            ),
+            hypot(
+              hypot(
+                2.0f*hypot(
+                  z_new[ind].s6,
+                  z_new[ind].s7
+                  ),
+                2.0f*hypot(
+                  z_new[ind].s8,
+                  z_new[ind].s9
+                  )
+                ),
+                2.0f*hypot(
+                  z_new[ind].sa,
+                  z_new[ind].sb
+                  )
+              )
+            )
+          );
     }
-    i = k*Nx*Ny*NUk+Nx*y + x;
+    fac *= alphainv;
+
     for (int uk=0; uk<NUk; uk++)
     {
-        if (fac > 1.0f) {z_new[i] /=fac;}
-        i+=Nx*Ny;
[…]
