### 2020-10-29 fcbd1aa
First Soft sense implementation
