+++ pyqmri/solver.py
@@ -18,6 +18,9 @@ import pyqmri.operator as operator
 from pyqmri._helper_fun import CLProgram as Program
 import sys
 import pyqmri.streaming as streaming
+
+from pyqmri._my_helper_fun.display_data import *
+
 DTYPE = np.complex64
 DTYPE_real = np.float32
 
@@ -425,6 +428,18 @@ class PDBaseSolver:
                     linops,
                     coils,
                     model)
+        elif reg_type == '':
+            pdop = PDSoftSenseSolver(
+                par,
+                irgn_par,
+                queue,
+                np.float32(1 / np.sqrt(8)),
+                init_fval,
+                prg,
+                linops,
+                coils,
+                model,
+                imagespace=imagespace)
         else:
             raise NotImplementedError
         return pdop
@@ -2308,3 +2323,2051 @@ class PDSolverStreamedTVSMS(PDSolverStreamedTV):
         lhs = np.sqrt(beta)*tau*np.abs(lhs1+lhs2+lhs3)**(1/2)
 
         return lhs, ynorm
+
+
+class PDSoftSenseBaseSolver:
+    """
+    Primal Dual splitting optimization.
+
+    This Class performs a primal-dual variable splitting based reconstruction
+    on single precission complex input data.
+    """
+
+    def __init__(self, par, irgn_par, queue, tau, fval, prg, coil, model):
+        """
+        PD reconstruction Object.
+
+        Args
+        ----
+          par (dict): A python dict containing the necessary information to
+            setup the object. Needs to contain the number of slices (NSlice),
+            number of scans (NScan), image dimensions (dimX, dimY), number of
+            coils (NC), sampling points (N) and read outs (NProj)
+            a PyOpenCL queue (queue) and the complex coil
+            sensitivities (C).
+          irgn_par (dict): A python dict containing the regularization
+            parameters for a given gauss newton step.
+          queue (list): A list of PyOpenCL queues to perform the optimization.
+          tau (float): Estimate of the initial step size based on the
+            operator norm of the linear operator.
+          fval (float): Estimate of the initial cost function value to
+            scale the displayed values.
+          prg (PyOpenCL Program): A PyOpenCL Program containing the
+            kernels for optimization.
+          reg_type (string): String to choose between "TV" and "TGV"
+            optimization.
+          data_operator (PyQMRI Operator): The operator to traverse from
+            parameter to data space.
+          coils (PyOpenCL Buffer or empty List): optional coil buffer.
+          NScan (int): Number of Scan which should be used internally. Do not
+            need to be the same number as in par["NScan"]
+          trafo (bool): Switch between radial (1) and Cartesian (0) fft.
+          and slice accelerated (1) reconstruction.
+        """
+        self.sigma = irgn_par["sigma"]
+        self.delta = irgn_par["delta"]
+        self.gamma = irgn_par["gamma"]
+        self.lambd = irgn_par["lambd"]
+        self.accelerated = irgn_par["accelerated"]
+        self.stag = irgn_par["stag"]
+        self.tol = irgn_par["tol"]
+        self.display_iterations = irgn_par["display_iterations"]
+        self.tau = tau
+        self.unknowns = par["unknowns"]
+        self.num_dev = len(par["num_dev"])
+        self.dz = par["dz"]
+        self._fval_init = fval
+        self._prg = prg
+        self._queue = queue
+        self._coils = coil
+        self.model = model
+
+    @staticmethod
+    def factory(
+            prg,
+            queue,
+            par,
+            irgn_par,
+            init_fval,
+            coils,
+            linops,
+            model=None,
+            reg_type='',
+            SMS=False,
+            streamed=False):
+        """
+        Generate a PDSolver object.
+
+        Args
+        ----
[…]
