+++ pyqmri/operator.py
@@ -276,6 +276,26 @@ class Operator(ABC):
                                            DTYPE_real)
         return op
 
+    @staticmethod
+    def SoftSenseOperatorFactory(par,
+                                prg,
+                                DTYPE,
+                                DTYPE_real,
+                                streamed=False):
+        """Sense forward/adjoint operator factory method."""
+        if streamed:
+            op = OperatorSoftSenseStreamed(par,
+                                           prg,
+                                           DTYPE,
+                                           DTYPE_real
+            )
+        else:
+            op = OperatorSoftSense(par,
+                                   prg[0],
+                                   DTYPE,
+                                   DTYPE_real)
+        return op
+
     def _defineoperator(self,
                         functions,
                         outp,
@@ -1942,3 +1962,396 @@ class OperatorFiniteSymGradientStreamed(Operator):
             self._ratio[idx].data,
             self.DTYPE_real(self._dz),
             wait_for=outp.events + inp[0].events + wait_for)
+
+
+class OperatorSoftSense(Operator):
+    """ Soft Sense Operator
+    """
+
+    def __init__(self, par, prg, DTYPE=np.complex64,
+                 DTYPE_real=np.float32, trafo=False):
+        super().__init__(par, prg, DTYPE, DTYPE_real)
+        self.queue = self.queue[0]
+        self.ctx = self.ctx[0]
+        self.NMaps = par["NMaps"]
+        self.NScan = 1
+        self.tmp_result = clarray.empty(
+            self.queue, (self.NScan, self.NC,
+                         self.NSlice, self.dimY, self.dimX),
+            self.DTYPE, "C")
+        if not trafo:
+            self.Nproj = self.dimY
+            self.N = self.dimX
+
+        self.NUFFT = CLnuFFT.create(self.ctx,
+                                    self.queue,
+                                    par,
+                                    radial=trafo,
+                                    DTYPE=DTYPE,
+                                    DTYPE_real=DTYPE_real)
+
+        self._check = np.ones((self.NScan, self.NC, self.NSlice, self.dimY, self.dimX), dtype=DTYPE_real)
+        self._check[..., ::2] = -1
+        self._check[..., 1::2, :] *= -1
+        if np.size(par["fft_dim"]) == 3:
+            self._check[..., 1::2, :, :] *= -1
+        self._check = clarray.to_device(self.queue, self._check)
+
+    def fwd(self, out, inp, wait_for=[]):
+        self.tmp_result.add_event(
+            self.prg.operator_fwd_ssense(
+                self.queue,
+                (self.NSlice, self.dimY, self.dimX),
+                None,
+                self.tmp_result.data,
+                inp[0].data,
+                inp[1].data,
+                np.int32(self.NC),
+                np.int32(self.NMaps),
+                wait_for=self.tmp_result.events + inp[0].events + wait_for))
+        self.tmp_result.add_event(
+            self.NUFFT.prg.masking(
+                self.queue,
+                (self.tmp_result.size,),
+                None,
+                self.tmp_result.data,
+                self._check.data,
+                wait_for=wait_for + self.tmp_result.events))
+        out.add_event(self.NUFFT.FFT(
+                out,
+                self.tmp_result,
+                wait_for=wait_for + out.events + self.tmp_result.events))
+        return self.NUFFT.prg.masking(
+            self.queue,
+            (out.size,),
+            None,
+            out.data,
+            self._check.data,
[…]

[…]
