+++ pyqmri/models/FFCPaperPhase.py
@@ -25,11 +25,13 @@ class Model(BaseModel):
 
         self.numT1Scale = len(self.b)-1
 
-        par["unknowns_TGV"] = 2 + len(self.b)
+        par["unknowns_TGV"] = 2 + len(self.b) + len(self.b) - 1
         par["unknowns_H1"] = 0
         par["unknowns"] = par["unknowns_TGV"]+par["unknowns_H1"]
 
         self.unknowns = par["unknowns"]
+        
+        t1min = np.min(self.t)/2
 
         for j in range(par["unknowns"]):
             self.uk_scale.append(1)
@@ -37,20 +39,25 @@ class Model(BaseModel):
         self.constraints.append(
             constraints(0,
                         1000,
-                        False))
-        self.constraints.append(
-            constraints(1,
-                        1,
                         True))
         self.constraints.append(
-            constraints(1e-10,
+            constraints(0,
+                        1,
+                        False))
+        self.constraints.append(
+            constraints(t1min,
                         2000,
                         True))
         for j in range(self.numT1Scale):
             self.constraints.append(
-                constraints(1e-10,
+                constraints(t1min,
                             2000,
                             True))
+        for j in range(self.numT1Scale):
+            self.constraints.append(
+                constraints(0,
+                            1,
+                            False))
         self._ind1 = 0
         self._ind2 = 0
         self._labels = []
@@ -67,6 +74,9 @@ class Model(BaseModel):
         for j in range(self.numT1Scale):
             tmp_x[3+j] *= self.uk_scale[3+j]
             ukname.append("T1_"+str(2+j))
+        for j in range(self.numT1Scale):
+            tmp_x[3+self.numT1Scale+j] *= self.uk_scale[3+self.numT1Scale+j]
+            ukname.append("al_"+str(2+j))
         const = []
         for constrained in self.constraints:
             const.append(constrained.real)
@@ -74,20 +84,14 @@ class Model(BaseModel):
                 "unknown_name": ukname,
                 "real_valued": const}
 
-    def _execute_forward_2D(self, x, islice):
-        pass
-
-    def _execute_gradient_2D(self, x, islice):
-        pass
-
     def _execute_forward_3D(self, x):
         S = np.zeros(
             (self.NScan, self.NSlice, self.dimY, self.dimX),
             dtype=self._DTYPE)
         t = self.t[0][:, None, None, None]
         S[:len(t)] = (
-            x[0] * self.uk_scale[0]
-            * (-x[1]*self.uk_scale[1]*self.b0 *
+            x[0] * self.uk_scale[0]*x[1]*self.uk_scale[1]
+            * (-self.b0 *
                np.exp(-t / (x[2] * self.uk_scale[2]))
                + (1 - np.exp(-t / (x[2] * self.uk_scale[2])))
                * self.b[0])
@@ -96,8 +100,8 @@ class Model(BaseModel):
             offset = len(self.t[j+1])
             t = self.t[j+1][:, None, None, None]
             S[offset*(j+1):offset*(j+2)] = (
-                x[0] * self.uk_scale[0]
-                * (-x[1]*self.uk_scale[1]*self.b0 *
+                x[0] * self.uk_scale[0]*x[3+self.numT1Scale+j]*self.uk_scale[3+self.numT1Scale+j]
+                * (-self.b0 *
[…]
