### 2020-11-02 e168e6f
Working on improved readme and examples
