+++ pyqmri/_helper_fun/_utils.py
@@ -63,11 +63,11 @@ def gen_default_config():
     config['TGV']["max_gn_it"] = '7'
     config['TGV']["lambd"] = '1e0'
     config['TGV']["gamma"] = '1e-3'
-    config['TGV']["delta"] = '1e-4'
+    config['TGV']["delta"] = '1e-2'
     config['TGV']["omega"] = '0'
     config['TGV']["display_iterations"] = '0'
     config['TGV']["gamma_min"] = '3e-4'
-    config['TGV']["delta_max"] = '1e-1'
+    config['TGV']["delta_max"] = '1e2'
     config['TGV']["omega_min"] = '0'
     config['TGV']["tol"] = '1e-6'
     config['TGV']["stag"] = '1e10'
@@ -81,11 +81,11 @@ def gen_default_config():
     config['TV']["max_gn_it"] = '7'
     config['TV']["lambd"] = '1e0'
     config['TV']["gamma"] = '1e-3'
-    config['TV']["delta"] = '1e-4'
+    config['TV']["delta"] = '1e-2'
     config['TV']["omega"] = '0'
     config['TV']["display_iterations"] = '0'
     config['TV']["gamma_min"] = '3e-4'
-    config['TV']["delta_max"] = '1e-1'
+    config['TV']["delta_max"] = '1e2'
     config['TV']["omega_min"] = '0'
     config['TV']["tol"] = '1e-6'
     config['TV']["stag"] = '1e10'
