### 2020-11-02 d9c9199
Small fixes for JOSS publication
