### 2020-11-02 b3ad660
Added a minimum required Sympy version (1.6.2
