### 2020-11-02 1374c9b
Added Sympy minimum version
