+++ requirements.txt
@@ -8,4 +8,4 @@ ipyparallel
 pyfftw
 pyqt5<5.13
 numexpr
-sympy
+sympy>=6.1.2
