+++ pyqmri/_helper_fun/_utils.py
@@ -61,36 +61,36 @@ def gen_default_config():
     config['TGV']["max_iters"] = '1000'
     config['TGV']["start_iters"] = '10'
     config['TGV']["max_gn_it"] = '7'
-    config['TGV']["lambd"] = '1e0'
-    config['TGV']["gamma"] = '1e-3'
-    config['TGV']["delta"] = '1e-4'
+    config['TGV']["lambd"] = '1e2'
+    config['TGV']["gamma"] = '5e-2'
+    config['TGV']["delta"] = '1e-1'
     config['TGV']["omega"] = '0'
     config['TGV']["display_iterations"] = '0'
-    config['TGV']["gamma_min"] = '3e-4'
-    config['TGV']["delta_max"] = '1e-1'
+    config['TGV']["gamma_min"] = '5e-3'
+    config['TGV']["delta_max"] = '1e2'
     config['TGV']["omega_min"] = '0'
     config['TGV']["tol"] = '1e-6'
     config['TGV']["stag"] = '1e10'
     config['TGV']["delta_inc"] = '10'
-    config['TGV']["gamma_dec"] = '0.7'
+    config['TGV']["gamma_dec"] = '0.5'
     config['TGV']["omega_dec"] = '0.5'
 
     config['TV'] = {}
     config['TV']["max_iters"] = '1000'
     config['TV']["start_iters"] = '10'
     config['TV']["max_gn_it"] = '7'
-    config['TV']["lambd"] = '1e0'
-    config['TV']["gamma"] = '1e-3'
-    config['TV']["delta"] = '1e-4'
+    config['TV']["lambd"] = '1e2'
+    config['TV']["gamma"] = '5e-2'
+    config['TV']["delta"] = '1e-1'
     config['TV']["omega"] = '0'
     config['TV']["display_iterations"] = '0'
-    config['TV']["gamma_min"] = '3e-4'
-    config['TV']["delta_max"] = '1e-1'
+    config['TV']["gamma_min"] = '5e-3'
+    config['TV']["delta_max"] = '1e2'
     config['TV']["omega_min"] = '0'
     config['TV']["tol"] = '1e-6'
     config['TV']["stag"] = '1e10'
     config['TV']["delta_inc"] = '10'
-    config['TV']["gamma_dec"] = '0.7'
+    config['TV']["gamma_dec"] = '0.5'
     config['TV']["omega_dec"] = '0.5'
 
     with open('default.ini', 'w') as configfile:
