### 2020-11-02 92adc88
Added CONTRIBUTING description
