+++ README.rst
@@ -149,7 +149,7 @@ or by fewing the documentation of pyqmri.pyqmri in python.
 
 If reconstructing fewer slices from the volume than acquired, slices will be picked symmetrically from the center of the volume. E.g. reconstructing only a single slice will reconstruct the center slice of the volume. 
 
-An example script, showing how to start and use PyQMRI can be found on _GoogleColab.
+An example script, showing how to use PyQMRI based on an exmaple VFA data set can be found on _GoogleColab.
 
 The config file (\*.ini):
 -------------------------   
@@ -213,5 +213,6 @@ at `[v0.1.0] <https://github.com/IMTtugraz/PyQMRI/tree/v.0.1.0>`_
 .. _zenodo: https://doi.org/10.5281/zenodo.1410918
 .. _NLINV: https://doi.org/10.1002/mrm.21691
 .. _PyOpenCL: https://github.com/inducer/pyopencl
+.. _GoogleColab: https://colab.research.google.com/drive/19BfSJmDPinZDY0m1sMAhETutIiJG3b33?usp=sharing
 .. _contributing: CONTRIBUTING.rst
 
