### 2020-11-02 #2727 @MaierOli2010
I've added basic Contributing guidelines to the project as well as a "Statement of Need" in the README.rst. Further, I've made a very basic "How to use" Guide on Google Colab which might needs more formatting but gives a potential use a quick insight into how to use PyQMRI.
