### 2020-11-02 e6635dd
Fixed typo in Sympy version
