### 2020-11-04 #2727 @DARSakthi
Hi everyone,

Apologies for the delay in responding to this review. October and November are very busy months -- some NIH deadlines to prepare for.

A few comments --

Firstly, the scholarly effort is substantial. I commend the authors on not only the code but the methods development.

I noticed you mention you added some google colab resources, but I can't seem to find them. Could you point me in their direction? Because google colab offers limited but free online GPU usage, I think it would be a good idea for you to move towards integrating it within your package. Perhaps as a separate release. This isn't crucial to publication -- merely a suggestion.

The documentation is sufficiently complete.

I recommend, prior to publication, some copy editing on the submitted manuscript. I can provide specific examples of sentences which I think could be better phrased later.

Otherwise I think this is a good paper. Once the above suggestion about language is addressed I would recommend acceptance for publication.

Cheers

Dalton Sakthivadivel
