### 2020-11-05 #2727 @MaierOli2010
Thank you for the comments!

The links i sin the "Running the reconstruction" section, last sentence. It is not very prominently placed.
You can also use [this link](https://colab.research.google.com/drive/19BfSJmDPinZDY0m1sMAhETutIiJG3b33?usp=sharing) to have a look.

If you could mark the sentences that should be rephrased that would be great.

Cheers

Oliver
