### 2020-11-06 927572b
added function to save images and some arguments to argument parser
