### 2020-11-09 #2727 @grlee77
I am also overall in favor of this publication and will update with any final comments later today.

Great job on the collab notebook. That will be very helpful to potential users.
