### 2020-11-09 #2727 @agahkarakuzu
@MaierOli2010 thank you sou much for providing the colab notebook! I was having some issues with the CUDA version I have on my local machine, this addition dispensed the need for dealing with it. Before I begin, I would like to give full disclosure that I am the author of [qMRLab](https://joss.theoj.org/papers/10.21105/joss.02343), a software for quantitative MRI analysis.

This submission effectively fulfills the scope eligibility, providing a powerful framework to ease the processing of 3D model-based reconstruction. I am happy to see the methods from [this work](https://onlinelibrary.wiley.com/doi/full/10.1002/mrm.27502) will be available as a Python package that meets JOSS standards. Besides the convenience `PyQMRI` brings to qMRI analysis with under-sampled data, the idea of tapping into symbolic expressions to introduce new models help users easily expand the toolbox according to their needs.

To be able to check all the boxes, I have some comments regarding the documentation and the article:

### Documentation

1) The `sphinx` auto-generated API documentation is of good service for users accustomed to the hierarchy of Python modules and for developers who’d like to contribute. Despite that it is delivered by a combination of `README.md` and the colab notebook, a nicely organized beginner’s guideline at the RTD page would make `PyQMRI` more user friendly.

The TOC in the colab notebook is intuitive, the example data is publicly available, and the basic functionality can be tested without worrying about the availability/compatibility of GPU resources. Centralizing the documentation by organizing these pieces together would lead to a more informative guideline and reduce the risk of users missing out on offered functionality. This way, `README.md` would become leaner, conveying the high-level functionality of the toolbox without overloading new users with usage details or the requisites about the input data structure.

2) Travis build checks the proper installation of the software and the minimum functionality. I see that the unit tests are performed by Jenkins given the need for GPU. Is it possible to show the state of the Jenkins build on GitHub, ideally with the build logs? I know such integrations are not provided out of the box by Jenkins, but if there is a plugin that can enable this through webhooks, it would do the finishing touch on the CI picture of the repository. If this is not possible, adding a small note about how these tests are performed in the README can be useful.

3) Community guidelines

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I was not able to locate the community guideline. Can you please point me to it if exists, or add it otherwise?

### Software paper

Overall, the article is well-written and clearly states which gap does the `PyQMRI` address.

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

However, I was not able to check this box as such information is currently missing in the article. For the sake of completeness, it is worth mentioning other open-source qMRI toolboxes (e.g. hMRI Toolbox, QUIT, qMRLab, QMAP, mrQ, to name a few) and where does `PyQMRI` stand in relation to them. I would like to note that two of these toolboxes are published in JOSS and I am looking forward to seeing PyQMRI added to that list.
