### 2020-11-09 #2727 @MaierOli2010
@agahkarakuzu thank you for the detailed list of comments and also for pointing me into the right direction, showing a multitude of qMRI toolboxes that are currently available. Unfortunately, I am quite busy this week and will be able to have a detailed look at them early next week. At the first glance the major difference, beside offering different types of models, seems to be that these expect image data as input and thus require image reconstruction algorithms prior to their usage (if undersample data is acquired of course).  Some of them integrate with registration toolboxes (e.g ITK, SPM) to offer registration prior to fitting.  I will have a closer look as soon as possible and update the paper with a corresponding section.

I will add the missing community guidlines as soon as possbile.

I am not sure about your point about the Jenkins build. The CI currently runs on a machine, nested in the university network. A quick search did not give any informaiton on how to upload build logs with webhooks to GitHub, only notifications about success/failure of test cases, which is already integrated in the Github repo. Might need to dig deeper into that but it might also not be possible at all without exposing the Jenkins host to the web, i.e. publicly hosting the jenkins webserver (which is an option I was thinking about in the past but it might also not be well received by our admin).

I do agree with the points about the documentation and will streamline and integrate the Google Colab parts to it. The parts about prerequest and specific requirements not covered by the notebook will move into the sphinx documentation and a link will be added to the `README.MD.`

Thank you again for your comments.

Best regards!
