### 2020-11-09 #2727 @agahkarakuzu
Thank you for the quick response @MaierOli2010, sure whenever you are available.

> which is already integrated in the Github repo

Sorry I was not clear, I meant the build status, like the one belongs to your Travis CI builds. At any rate, this is not a major issue, it would be nice to have public access to that information. But if the Jenkins instance's being installed on a campus server is going to make it tricky and/or if this is not an easy task, feel free to skip this one.
