### 2020-11-10 #2727 @grlee77
> I was not able to locate the community guideline. Can you please point me to it if exists, or add it otherwise?

I considered the `CONTRIBUTING.rst` file to fill this role. It doesn't cover things like expected code of conduct/behavior, but I don't think that is necessarily a requirement? If interested, feel free to adapt a basic, informal set of community guidelines like [this one from scikit-image](https://scikit-image.org/community_guidelines.html).
