### 2020-11-10 #2727 @grlee77
The API documentation seems fine and I think the new collab notebook serves as a reasonable beginner-friendly introduction to usage of the software so I have now checked off these items in the checklist.

I agree with the point related to a need for a brief summary of the state of the field (and corresponding references) which is why these are the only remaining items to address for me. I agree with @MaierOli2010 that this package is fairly unique in that reconstruction and quantification are combined in a single optimization problem from the raw k-space data. I think [BART](https://mrirecon.github.io/bart/) has something similar for non-linear inverse reconstruction for T1 mapping, but that is the only one I can recall offhand. I think the model in BART is hard-coded, though, so the user cannot easily modify it.
