### 2020-11-10 ccc435d
Removed DCF for testing. Further finetuning of FFC
