+++ pyqmri/models/FFCPaper.py
@@ -335,7 +335,7 @@ class Model(BaseModel):
         self.dscale = args[1]
         self.images = np.reshape(np.abs(args[0]/args[1]),
                                  self.t.shape+args[0].shape[-3:])
-        test_M0 = self.dscale*np.ones(
+        test_M0 = 1*np.ones(
             (self.NSlice, self.dimY, self.dimX), dtype=self._DTYPE)
         self.constraints[0].update(1/args[1])
         test_Xi = 1*np.ones(
