### 2020-11-10 792a7f2
Reduced sice of unittests on GPU to avoid out of memory errors
