### 2020-11-10 6ea4684
Updated Unittests from JOSS-pub
