### 2020-11-10 0245235
Merging unittests and new Jenkinsfile from JOSS_pub
