+++ requirements.txt
@@ -9,3 +9,4 @@ pyfftw
 pyqt5<5.13
 numexpr
 sympy>=1.6.2
+nbsphinx
