+++ doc/api.rst
@@ -1,0 +1,10 @@
+API Documentation
+=================
+.. toctree::
+   ./_modules/pyqmri
+   ./_modules/solver
+   ./_modules/streaming
+   ./_modules/operator
+   ./_modules/models
+   ./_modules/irgn
+   ./_modules/transforms
