### 2020-11-10 #2727 @DARSakthi
Hello,

The colab notebook works wonderfully. I will state again that I think it will attract a wider user base -- even in this era of high powered computing, not everyone has consistent access to a GPU.

As far as language goes, for example, the very first sentence is a bit messy. I would try and make it flow better. I recommend a copy editing software like Grammarly or the like. I have marked some portions of it [here](https://drive.google.com/file/d/18n6jr35B2_qpqIBY6cP56y9hynVyjrUr/view?usp=sharing) with some possible revisions. Do let me know if you can't access, or would like me to transcribe the comments as a reply in this thread. Overall they are minor suggestions, and _only_ suggestions -- you must write your paper the way you see fit. The algorithmic section was notably well written.

Regarding other reviewer comments, I found the description of the 'state of the field' sufficiently discussed the current standards in qMR -- but I do agree with suggestions for giving specific examples of similar packages.
