### 2020-11-10 #2727 @DARSakthi
By the way -- Microsoft word has a pretty nice grammar and spelling processor these days. This could provide some guidance. Especially if you don't want to spend the money on a grammarly license, as I believe it has a cost associated. Paper is attached in case of access issues.
[qmri_joss.pdf](https://github.com/openjournals/joss-reviews/files/5519776/qmri_joss.pdf)
