### 2020-11-11 b1a92dc
Streamlined README
