+++ README.rst
@@ -11,166 +11,24 @@ PyQMRI - Model-Based Parameter Quantification
 PyQMRI is a Python based toolbox for quantitative Magnetic Resonance Imaging (MRI). Utilizing _PyOpenCL and a double-buffering scheme, 
 it enables the accelerated reconsruction and fitting of arbitrary large datasets on memory limited GPUs.
 Currently, PyQMRI supports the processing of 3D Cartesian and non-Cartesian (stack-of-) data.
-In addition to the currently existing models, new models can be added using a simple text-based model-file.
 
-Prerequests:
-------------------------
-* Requires OpenCL_ >= 1.2
-* Requires clfft_
-* Requires gpyfft_
-* Requires Cython, pip >= 19, python >= 3.6
+Examples include T1 quantification from variable flip angle or 
+inversion-recovery Look-Locker data, T2 quantification using a 
+mono-exponential fit, or Diffusion Tensor quantification. 
 
-The Software is tested on Linux using the latest Nvidia driver (418.56 CUDA Version: 10.1) but should be compatible with older drivers as well as different hardware (AMD). The following Installation Guide is targeted at Ubuntu but should work on any distribution provided the required packages are present (could be differently named).
+In addition, a Genereal Model exists that can be invoced 
+using a text file containing the analytical signal equation.
 
-* It is highly recommended to use an Anaconda environment
+For a real world usage example have a look at the `Quickstart Guide`_.
+The example can also be run interactively using GoogleColab_.
 
-.. role:: bash(code)
-   :language: bash
-   
-.. role:: python(code)
-   :language: python
-   
-   
-Quick Installing Guide:
-------------------------
-First make sure that you have a working OpenCL installation
-
-* OpenCL is usually shipped with GPU driver (Nvidia/AMD)
-* Install the ocl_icd and the OpenCL-Headers
-
-  :bash:`apt-get install ocl_icd* opencl-headers`
-    
-Possible restart of system after installing new drivers and check if OpenCL is working
-
-* Build clinfo_:
-* Run clinfo_ in terminal and check for errors
-
-Install clFFT library:  
-
-* Either use the package repository,e.g.:
-
-  :bash:`apt-get install libclfft*`
-
-* Or download a prebuild binary of clfft_
-
-  - Please refer to the clfft_ docs regarding building
-  - If build from source symlink clfft_ libraries from lib64 to the lib folder and run :bash:`ldconfig`
-    
-Install gpyfft_ by following the instruction on the GitHub page. 
-  
-To Install PyQMRI, a simple
-  
-:bash:`pip install pyqmri`
-    
-should be sufficient to install the latest release.
-    
-Alternatively, clone the git repository and navigate to the root directory of PyQMRI. Typing
-  
-:bash:`pip install .`
-    
-should take care of the other dependencies using PyPI and install the package. 
-     
-In case OCL > 1.2 is present, e.g. by some CPU driver, and NVidia GPUs needs to be used the flag
-PRETENED_OCL 1.2 has to be passed to PyOpenCL during the build process. This 
-can be done by:
-
-.. code-block:: bash
-
-    ./configure.py --cl-pretend-version=1.2
-    rm -Rf build
-    python setup.py install
+Installation and usage guides, as well as API documentaiton, can be found in the Documentation_
 
 
 Sample Data
 -----------
 In-vivo datasets used in the original publication (doi: `[10.1002/mrm.27502]`_) can be found at zenodo_. If you use the sample data with the recent release of PyQMRI please delete the "Coils"
 entry in the .h5 to force a recomputation of the receive coil sensitivities as the orientation does not match the data.
-
-Prerequests on the .h5 file:
[…]
