### 2020-11-11 4ed6904
Reworked README and included CodeOfConduct
