### 2020-11-12 b079d1b
added function for initial cost calculation; fixing input shape for dual update with TGV; fixed lambda to 1 for PD algo without regularization; added function for step size estimation with power iterations
