+++ main.py
@@ -59,7 +59,90 @@ def _setup_irgn_par(irgn_par, myargs):
     irgn_par["gamma"] = 1
 
 
+def _power_iterations(par, x, cmaps, op, iters=100):
+    queue = par["queue"][0]
+    x = x.astype(DTYPE)
+    x = cla.to_device(queue, x)
+
+    if len(cmaps) > 0:
+        cmaps = cmaps.astype(DTYPE)
+        cmaps = cla.to_device(queue, cmaps)
+        y = op.adjoop([op.fwdoop([x, cmaps]), cmaps]).get()
+    else:
+        y = op.adjoop(op.fwdoop(x)).get()
+
+    for i in range(iters):
+        y_norm = np.linalg.norm(y)
+        x = y / y_norm if y_norm != 0 else y
+        x = cla.to_device(queue, x)
+        y = op.adjoop([op.fwdoop([x, cmaps]), cmaps]).get() if len(cmaps) > 0 else op.adjoop(op.fwdoop(x)).get()
+        x = x.get()
+        l1 = np.vdot(y, x)
+
+    return np.sqrt(np.max(np.abs(l1)))
+
+
+def _calc_step_size(args, par, ksp, cmaps):
+
+    if DTYPE == np.complex128:
+        file = open(
+            resource_filename(
+                'pyqmri', 'kernels/OpenCL_Kernels_double.c'))
+    else:
+        file = open(
+            resource_filename(
+                'pyqmri', 'kernels/OpenCL_Kernels.c'))
+    prg = Program(
+        par["ctx"][0],
+        file.read())
+    file.close()
+
+    op = pyqmirop.OperatorSoftSense(par, prg)
+    grad_op = pyqmirop.OperatorFiniteGradient(par, prg)
+    symgrad_op = pyqmirop.OperatorFiniteSymGradient(par, prg)
+
+    x = np.random.randn(1, par["NSlice"], par["dimY"], par["dimX"]) + \
+        1j * np.random.randn(1, par["NSlice"], par["dimY"], par["dimX"])
+
+    opnorm_1 = _power_iterations(par, x, cmaps[0], op)
+
+    x = np.random.randn(1, par["NSlice"], par["dimY"], par["dimX"]) + \
+        1j * np.random.randn(1, par["NSlice"], par["dimY"], par["dimX"])
+
+    opnorm_2 = _power_iterations(par, x, cmaps[1], op)
+
+    opnorm_grad = _power_iterations(par, x, [], grad_op)
+
+    x_symgrad = np.random.randn(1, par["NSlice"], par["dimY"], par["dimX"], 4) + \
+        1j * np.random.randn(1, par["NSlice"], par["dimY"], par["dimX"], 4)
+
+    opnorm_symgrad = _power_iterations(par, x_symgrad, [], symgrad_op)
+
+    K_ssense = np.array([opnorm_1, opnorm_2])
+    K_ssense_tv = np.array([[opnorm_1,      opnorm_2],
+                            [opnorm_grad,   0],
+                            [0,             opnorm_grad]])
+
+    K_ssense_tgv = np.array([[opnorm_1,     opnorm_2,       0,              0],
+                             [opnorm_grad,  0,              1,              0],
+                             [0,            0,              opnorm_symgrad, 0],
+                             [0,            opnorm_grad,    0,              1],
+                             [0,            0,              0,              opnorm_symgrad]])
+
+    if args.reg_type == '':
+        tau = 1 / np.sqrt(np.vdot(K_ssense, K_ssense))
[…]
