### 2020-11-12 9457a96
set argument lambda in argument parser to type float
