### 2020-11-13 c0946e6
Reworking Paper
