+++ paper.bib
@@ -288,3 +288,129 @@ url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.20283},
 volume = {52},
 year = {2004}
 }
+
+@article{hMRI,
+abstract = {Neuroscience and clinical researchers are increasingly interested in quantitative magnetic resonance imaging (qMRI) due to its sensitivity to micro-structural properties of brain tissue such as axon, myelin, iron and water concentration. We introduce the hMRI-toolbox, an open-source, easy-to-use tool available on GitHub, for qMRI data handling and processing, presented together with a tutorial and example dataset. This toolbox allows the estimation of high-quality multi-parameter qMRI maps (longitudinal and effective transverse relaxation rates R1 and R2⋆, proton density PD and magnetisation transfer MT saturation) that can be used for quantitative parameter analysis and accurate delineation of subcortical brain structures. The qMRI maps generated by the toolbox are key input parameters for biophysical models designed to estimate tissue microstructure properties such as the MR g-ratio and to derive standard and novel MRI biomarkers. Thus, the current version of the toolbox is a first step towards in vivo histology using MRI (hMRI) and is being extended further in this direction. Embedded in the Statistical Parametric Mapping (SPM) framework, it benefits from the extensive range of established SPM tools for high-accuracy spatial registration and statistical inferences and can be readily combined with existing SPM toolboxes for estimating diffusion MRI parameter maps. From a user's perspective, the hMRI-toolbox is an efficient, robust and simple framework for investigating qMRI data in neuroscience and clinical research.},
+author = {Tabelow, Karsten and Balteau, Evelyne and Ashburner, John and Callaghan, Martina F and Draganski, Bogdan and Helms, Gunther and Kherif, Ferath and Leutritz, Tobias and Lutti, Antoine and Phillips, Christophe and Reimer, Enrico and Ruthotto, Lars and Seif, Maryam and Weiskopf, Nikolaus and Ziegler, Gabriel and Mohammadi, Siawoosh},
+doi = {https://doi.org/10.1016/j.neuroimage.2019.01.029},
+issn = {1053-8119},
+journal = {Neuroimage},
+keywords = { In vivo histology, Microstructure, Multi-parameter mapping, Relaxometry, SPM toolbox,Quantitative MRI},
+pages = {191--210},
+title = {{hMRI – A toolbox for quantitative MRI in neuroscience and clinical research}},
+url = {http://www.sciencedirect.com/science/article/pii/S1053811919300291},
+volume = {194},
+year = {2019}
+}
+
+@article{QUIT,
+  doi = {10.21105/joss.00656},
+  url = {https://doi.org/10.21105/joss.00656},
+  year = {2018},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {26},
+  pages = {656},
+  author = {Tobias C. Wood},
+  title = {QUIT: QUantitative Imaging Tools},
+  journal = {Journal of Open Source Software}
+}
+
+
+@article{qMRLab,
+  doi = {10.21105/joss.02343},
+  url = {https://doi.org/10.21105/joss.02343},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {53},
+  pages = {2343},
+  author = {Agah Karakuzu and Mathieu Boudreau and Tanguy Duval and Tommy Boshkovski and Ilana R. Leppert and Jean-François Cabana and Ian Gagnon and Pascale Beliveau and G. Bruce Pike and Julien Cohen-Adad and Nikola Stikov},
[…]
