### 2020-11-16 #2727 @MaierOli2010
@agahkarakuzu  I've finally managed to get a working integration with Jenkins through webhooks. There is just one downside to this, as soon as I have to restart the web forwarding the hook changes and the README needs to be updated with the new links. Other than that it is possible to view the Jenkins build and all the reports now by clicking on the banner. One small thing that is currently not linking correctly is the "details" link if clicking on the green/red mark next to the commit which still points to localhost.
