### 2020-11-16 #2727 @grlee77
Thanks @MaierOli2010, , the new text gives a good overview of related tools. I have checked off the remaining items under my review above, but the new text does need some additional proofreading/typo fixes:

specific typos:
- apllcation -> application
- "PyQMRIis the only available"  (missing space after PyQMRI)
- qunatification -> quantification

reference formatting issues:
- The reference entries for Friston2007, Tabelow2019 and Wang2018 have duplicated https://doi.org
- The reference for Sumpf2012 looks incomplete
