### 2020-11-16 350bd59
Update paper.bib
Removed duplicated https://doi.org
