+++ paper.bib
@@ -292,7 +292,7 @@ year = {2004}
 @article{hMRI,
 abstract = {Neuroscience and clinical researchers are increasingly interested in quantitative magnetic resonance imaging (qMRI) due to its sensitivity to micro-structural properties of brain tissue such as axon, myelin, iron and water concentration. We introduce the hMRI-toolbox, an open-source, easy-to-use tool available on GitHub, for qMRI data handling and processing, presented together with a tutorial and example dataset. This toolbox allows the estimation of high-quality multi-parameter qMRI maps (longitudinal and effective transverse relaxation rates R1 and R2⋆, proton density PD and magnetisation transfer MT saturation) that can be used for quantitative parameter analysis and accurate delineation of subcortical brain structures. The qMRI maps generated by the toolbox are key input parameters for biophysical models designed to estimate tissue microstructure properties such as the MR g-ratio and to derive standard and novel MRI biomarkers. Thus, the current version of the toolbox is a first step towards in vivo histology using MRI (hMRI) and is being extended further in this direction. Embedded in the Statistical Parametric Mapping (SPM) framework, it benefits from the extensive range of established SPM tools for high-accuracy spatial registration and statistical inferences and can be readily combined with existing SPM toolboxes for estimating diffusion MRI parameter maps. From a user's perspective, the hMRI-toolbox is an efficient, robust and simple framework for investigating qMRI data in neuroscience and clinical research.},
 author = {Tabelow, Karsten and Balteau, Evelyne and Ashburner, John and Callaghan, Martina F and Draganski, Bogdan and Helms, Gunther and Kherif, Ferath and Leutritz, Tobias and Lutti, Antoine and Phillips, Christophe and Reimer, Enrico and Ruthotto, Lars and Seif, Maryam and Weiskopf, Nikolaus and Ziegler, Gabriel and Mohammadi, Siawoosh},
-doi = {https://doi.org/10.1016/j.neuroimage.2019.01.029},
+doi = {10.1016/j.neuroimage.2019.01.029},
 issn = {1053-8119},
 journal = {Neuroimage},
 keywords = { In vivo histology, Microstructure, Multi-parameter mapping, Relaxometry, SPM toolbox,Quantitative MRI},
@@ -373,7 +373,7 @@ pages={1667–1672}}
 address = {London},
 author = {Friston, K},
 booktitle = {Stat. Parametr. Mapp.},
-doi = {https://doi.org/10.1016/B978-012372560-8/50002-4},
+doi = {10.1016/B978-012372560-8/50002-4},
 editor = {FRISTON, KARL and ASHBURNER, JOHN and KIEBEL, STEFAN and NICHOLS, THOMAS and PENNY, WILLIAM},
 isbn = {978-0-12-372560-8},
 pages = {10--31},
@@ -408,7 +408,7 @@ volume = {79},
 number = {2},
 pages = {730-740},
 keywords = {model-based reconstruction, sparsity constraint, Look-Locker, T1 mapping; parallel imaging},
-doi = {https://doi.org/10.1002/mrm.26726},
+doi = {10.1002/mrm.26726},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.26726},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/mrm.26726},
[…]
