+++ pyqmri/operator.py
@@ -103,6 +103,10 @@ class Operator(ABC):
         self.DTYPE = DTYPE
         self.DTYPE_real = DTYPE_real
         self.par_slices = self.NSlice
+        self._unknown_shape = (self.unknowns, 
+                               self.NSlice, 
+                               self.dimY, 
+                               self.dimX)
         self._overlap = 0
 
     @abstractmethod
@@ -465,10 +469,10 @@ class OperatorImagespace(Operator):
         else:
             wait_for = []
         return self.prg.operator_fwd_imagespace(
-            self.queue, (self.NSlice, self.dimY, self.dimX), None,
+            self.queue, (np.prod((np.prod((self.NSlice, self.dimY, self.dimX)),)),), None,
             out.data, inp[0].data, inp[2].data,
             np.int32(self.NScan),
-            np.int32(self.unknowns),
+            *np.int32(self._unknown_shape),
             wait_for=inp[0].events + out.events + wait_for)
 
     def fwdoop(self, inp, **kwargs):
@@ -500,10 +504,10 @@ class OperatorImagespace(Operator):
             self.queue, (self.NScan, self.NSlice, self.dimY, self.dimX),
             self.DTYPE, "C")
         tmp_result.add_event(self.prg.operator_fwd_imagespace(
-            self.queue, (self.NSlice, self.dimY, self.dimX), None,
+            self.queue, (np.prod((np.prod((self.NSlice, self.dimY, self.dimX)),)),), None,
             tmp_result.data, inp[0].data, inp[2].data,
             np.int32(self.NScan),
-            np.int32(self.unknowns),
+            *np.int32(self._unknown_shape),
             wait_for=inp[0].events + wait_for))
         return tmp_result
 
@@ -533,10 +537,10 @@ class OperatorImagespace(Operator):
         else:
             wait_for = []
         return self.prg.operator_ad_imagespace(
-            out.queue, (self.NSlice, self.dimY, self.dimX), None,
+            out.queue, (np.prod((self.NSlice, self.dimY, self.dimX)),), None,
             out.data, inp[0].data, inp[2].data,
             np.int32(self.NScan),
-            np.int32(self.unknowns),
+            *np.int32(self._unknown_shape),
             wait_for=wait_for + inp[0].events + out.events)
 
     def adjoop(self, inp, **kwargs):
@@ -568,10 +572,10 @@ class OperatorImagespace(Operator):
             self.queue, (self.unknowns, self.NSlice, self.dimY, self.dimX),
             dtype=self.DTYPE)
         self.prg.operator_ad_imagespace(
-            out.queue, (self.NSlice, self.dimY, self.dimX), None,
+            out.queue, (np.prod((self.NSlice, self.dimY, self.dimX)),), None,
             out.data, inp[0].data, inp[2].data,
             np.int32(self.NScan),
-            np.int32(self.unknowns),
+            *np.int32(self._unknown_shape),
             wait_for=wait_for + inp[0].events + out.events).wait()
         return out
 
@@ -600,11 +604,11 @@ class OperatorImagespace(Operator):
         else:
             wait_for = []
         return self.prg.update_Kyk1_imagespace(
-            self.queue, (self.NSlice, self.dimY, self.dimX), None,
+            self.queue, (np.prod((self.NSlice, self.dimY, self.dimX)),), None,
             out.data, inp[0].data, inp[3].data, inp[1].data,
             np.int32(self.NScan),
             inp[4].data,
-            np.int32(self.unknowns),
+            *np.int32(out.shape),
             self.DTYPE_real(self._dz),
             wait_for=(inp[0].events + out.events
                       + inp[1].events + wait_for))
@@ -695,13 +699,13 @@ class OperatorKspace(Operator):
         self._tmp_result.add_event(
[…]
