+++ pyqmri/models/Ivim.py
@@ -1,0 +1,337 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+from pyqmri.models.template import BaseModel, constraints, DTYPE
+import matplotlib.gridspec as gridspec
+import matplotlib.pyplot as plt
+import numpy as np
+plt.ion()
+
+import ipdb
+
+
+class Model(BaseModel):
+    def __init__(self, par):
+        super().__init__(par)
+        self.NSlice = par['NSlice']
+
+        self.figure_phase = None
+
+        self.b = np.ones((self.NScan, 1, 1, 1))
+        self.dir = par["DWI_dir"].T
+        for i in range(self.NScan):
+            self.b[i, ...] = par["b_value"][i] * np.ones((1, 1, 1))
+
+        if np.max(self.b) > 100:
+            self.b /= 1000
+
+        self.dir = self.dir[:, None, None, None, :]
+        par["unknowns_TGV"] = 4
+        par["unknowns_H1"] = 0
+        par["unknowns"] = par["unknowns_TGV"] + par["unknowns_H1"]
+        self.unknowns = par["unknowns_TGV"] + par["unknowns_H1"]
+        self.uk_scale = []
+        for j in range(self.unknowns):
+            self.uk_scale.append(1)
+
+        try:
+            self.b0 = np.flip(
+                np.transpose(par["file"]["b0"][()], (0, 2, 1)), 0)
+        except KeyError:
+            print("No b0 image provided")
+            self.b0 = None
+
+        self.constraints.append(
+            constraints(
+                0 / self.uk_scale[0],
+                10 / self.uk_scale[0],
+                False))
+        self.constraints.append(
+            constraints(
+                (-0 / self.uk_scale[1]),
+                (10e0 / self.uk_scale[1]),
+                True))
+        self.constraints.append(
+            constraints(
+                (0 / self.uk_scale[2]),
+                (1 / self.uk_scale[2]),
+                True))
+        self.constraints.append(
+            constraints(
+                (0 / self.uk_scale[3]),
+                (150 / self.uk_scale[3]),
+                True))
+
+    def rescale(self, x):
+        M0 = x[0, ...] * self.uk_scale[0]
+        ADC = x[1, ...] * self.uk_scale[1]
+        f = x[2, ...] * self.uk_scale[2]
+        ADC_ivim = x[3, ...] * self.uk_scale[3]
+
+        return np.array((M0, ADC,
+                         f, ADC_ivim))
+
+    def _execute_forward_2D(self, x, islice):
+        print("2D Functions not implemented")
+        raise NotImplementedError
+
+    def _execute_gradient_2D(self, x, islice):
+        print("2D Functions not implemented")
+        raise NotImplementedError
+
+    def _execute_forward_3D(self, x):
+        #ipdb.set_trace()
+        
+        ADC = x[1, ...] * self.uk_scale[1]
+
+        S = (x[0, ...] * self.uk_scale[0] * (
+                x[2, ...] * self.uk_scale[2]
+                * np.exp(-(x[3, ...] * self.uk_scale[3]) * self.b)
+                + (1-x[2, ...] * self.uk_scale[2])
+                * np.exp(- ADC * self.b)
+             )).astype(DTYPE)
+
+        S *= self.phase
+        S[~np.isfinite(S)] = 0
+        return S
+
+    def _execute_gradient_3D(self, x):
+        ADC = x[1, ...] * self.uk_scale[1]
+
+        grad_M0 = self.uk_scale[0] * (
+            x[2, ...] * self.uk_scale[2]
+            * np.exp(- (x[3, ...] * self.uk_scale[3]) * self.b)
[…]
