+++ pyqmri/models/Diffivim_Dstar.py
@@ -1,0 +1,589 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+from pyqmri.models.template import BaseModel, constraints, DTYPE
+import matplotlib.gridspec as gridspec
+import matplotlib.pyplot as plt
+import numpy as np
+plt.ion()
+
+
+class Model(BaseModel):
+    def __init__(self, par):
+        super().__init__(par)
+        self.NSlice = par['NSlice']
+
+        self.figure_phase = None
+
+        self.b = np.ones((self.NScan, 1, 1, 1))
+        self.dir = par["DWI_dir"].T
+        for i in range(self.NScan):
+            self.b[i, ...] = par["b_value"][i] * np.ones((1, 1, 1))
+
+        if np.max(self.b) > 100:
+            self.b /= 1000
+
+        self.dir = self.dir[:, None, None, None, :]
+        par["unknowns_TGV"] = 9
+        par["unknowns_H1"] = 0
+        par["unknowns"] = par["unknowns_TGV"] + par["unknowns_H1"]
+        self.unknowns = par["unknowns_TGV"] + par["unknowns_H1"]
+        self.uk_scale = []
+        for j in range(self.unknowns):
+            self.uk_scale.append(1)
+
+        try:
+            self.b0 = np.flip(
+                np.transpose(par["file"]["b0"][()], (0, 2, 1)), 0)
+        except KeyError:
+            print("No b0 image provided")
+            self.b0 = None
+
+        self.constraints.append(
+            constraints(
+                0 / self.uk_scale[0],
+                10 / self.uk_scale[0],
+                False))
+        self.constraints.append(
+            constraints(
+                (-10e0 / self.uk_scale[1]),
+                (10e0 / self.uk_scale[1]),
+                True))
+        self.constraints.append(
+            constraints(
+                (-10e0 / self.uk_scale[2]),
+                (10e0 / self.uk_scale[2]),
+                True))
+        self.constraints.append(
+            constraints(
+                (-10e0 / self.uk_scale[3]),
+                (10e0 / self.uk_scale[3]),
+                True))
+        self.constraints.append(
+            constraints(
+                (-10e0 / self.uk_scale[4]),
+                (10e0 / self.uk_scale[4]),
+                True))
+        self.constraints.append(
+            constraints(
+                (-10e0 / self.uk_scale[5]),
+                (10e0 / self.uk_scale[5]),
+                True))
+        self.constraints.append(
+            constraints(
+                (-10e0 / self.uk_scale[6]),
+                (10e0 / self.uk_scale[6]),
+                True))
+        self.constraints.append(
+            constraints(
+                (0 / self.uk_scale[7]),
+                (1 / self.uk_scale[7]),
+                True))
+        # loosen D* constraint
+        self.constraints.append(
+            constraints(
+                (0 / self.uk_scale[8]),
+                (150 / self.uk_scale[8]),
+                True))
+
+    def rescale(self, x):
+        M0 = x[0, ...] * self.uk_scale[0]
+        ADC_x = (np.real(x[1, ...]**2) * self.uk_scale[1]**2)
+        ADC_xy = (np.real(x[2, ...] * self.uk_scale[2] *
+                          x[1, ...] * self.uk_scale[1]))
+        ADC_y = (np.real(x[2, ...]**2 * self.uk_scale[2]**2 +
+                         x[3, ...]**2 * self.uk_scale[3]**2))
[…]
