### 2020-11-18 fffabe2
Update pyqmri.py
