### 2020-11-18 68c8f58
Update Diffivim_Dstar.py
