+++ pyqmri/models/Diffivim_Dstar.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-from pyqmri.models.template import BaseModel, constraints, DTYPE
+from pyqmri.models.template import BaseModel, constraints
 import matplotlib.gridspec as gridspec
 import matplotlib.pyplot as plt
 import numpy as np
@@ -139,7 +139,7 @@ class Model(BaseModel):
                 * np.exp(-(x[8, ...] * self.uk_scale[8]) * self.b)
                 + (1-x[7, ...] * self.uk_scale[7])
                 * np.exp(- ADC * self.b)
-             )).astype(DTYPE)
+             )).astype(self.DTYPE)
 
         S *= self.phase
         S[~np.isfinite(S)] = 0
@@ -223,7 +223,7 @@ class Model(BaseModel):
              grad_ADC_z,
              grad_ADC_yz,
              grad_f,
-             grad_ADC_ivim], dtype=DTYPE)
+             grad_ADC_ivim], dtype=self.DTYPE)
         grad[~np.isfinite(grad)] = 0
         grad *= self.phase
         return grad
@@ -569,10 +569,10 @@ class Model(BaseModel):
             test_M0 = self.b0
         else:
             test_M0 = args[0][0]
-        ADC = 1 * np.ones(args[0].shape[-3:], dtype=DTYPE)
+        ADC = 1 * np.ones(args[0].shape[-3:], dtype=self.DTYPE)
         # change initial guess for f, D*
-        f = 0.1 * np.ones(args[0].shape[-3:], dtype=DTYPE)
-        ADC_ivim = 10 * np.ones(args[0].shape[-3:], dtype=DTYPE)
+        f = 0.1 * np.ones(args[0].shape[-3:], dtype=self.DTYPE)
+        ADC_ivim = 10 * np.ones(args[0].shape[-3:], dtype=self.DTYPE)
 
         x = np.array(
                 [
@@ -585,5 +585,5 @@ class Model(BaseModel):
                     0 * ADC,
                     f,
                     ADC_ivim],
-                dtype=DTYPE)
+                dtype=self.DTYPE)
         self.guess = x
