### 2020-11-18 3393f16
Update Diffivim_Dstar.py
