+++ pyqmri/models/Diffivim_Dstar.py
@@ -139,7 +139,7 @@ class Model(BaseModel):
                 * np.exp(-(x[8, ...] * self.uk_scale[8]) * self.b)
                 + (1-x[7, ...] * self.uk_scale[7])
                 * np.exp(- ADC * self.b)
-             )).astype(self.DTYPE)
+             )).astype(self._DTYPE)
 
         S *= self.phase
         S[~np.isfinite(S)] = 0
@@ -223,7 +223,7 @@ class Model(BaseModel):
              grad_ADC_z,
              grad_ADC_yz,
              grad_f,
-             grad_ADC_ivim], dtype=self.DTYPE)
+             grad_ADC_ivim], dtype=self._DTYPE)
         grad[~np.isfinite(grad)] = 0
         grad *= self.phase
         return grad
@@ -569,10 +569,10 @@ class Model(BaseModel):
             test_M0 = self.b0
         else:
             test_M0 = args[0][0]
-        ADC = 1 * np.ones(args[0].shape[-3:], dtype=self.DTYPE)
+        ADC = 1 * np.ones(args[0].shape[-3:], dtype=self._DTYPE)
         # change initial guess for f, D*
-        f = 0.1 * np.ones(args[0].shape[-3:], dtype=self.DTYPE)
-        ADC_ivim = 10 * np.ones(args[0].shape[-3:], dtype=self.DTYPE)
+        f = 0.1 * np.ones(args[0].shape[-3:], dtype=self._DTYPE)
+        ADC_ivim = 10 * np.ones(args[0].shape[-3:], dtype=self._DTYPE)
 
         x = np.array(
                 [
@@ -585,5 +585,5 @@ class Model(BaseModel):
                     0 * ADC,
                     f,
                     ADC_ivim],
-                dtype=self.DTYPE)
+                dtype=self._DTYPE)
         self.guess = x
