### 2020-11-19 d86bfbf
Update Ivim.py
bug fix Ivim.py
