### 2020-11-19 1c7d320
Update Diffivim_Dstar.py
bug fix Diffivim_Dstar.py
