### 2020-11-20 #2727 @MaierOli2010
Due to security concerns I've temporarily disabled the link to the Jenkins server. Working with our IT Security to get an official solution but it might take some time.
