### 2020-11-22 ead9aca
Update paper.bib
Additional information for Sumpf 2012
