+++ paper.bib
@@ -193,12 +193,14 @@ url = {http://link.springer.com/10.1007/978-3-642-54774-4{\_}3},
 year = {2014}
 }
 
-@article{Sumpf2012,
+@phdthesis{Sumpf2012,
 author = {Sumpf, Tilman},
 file = {:mnt/data/Papers/ASL/ASL.bib:bib},
 keywords = {model-based reconstruction undersampling MRI T2 re},
 title = {{Model-based T 2 Relaxometry using Undersampled Magnetic Resonance Imaging}},
-year = {2012}
+school = {{Technical University of Braunschweig}},
+year = {2012},
+doi = {10.24355/dbbs.084-201305311128-0}
 }
 
 @article{Roeloffs2016,
