+++ paper.bib
@@ -193,14 +193,30 @@ url = {http://link.springer.com/10.1007/978-3-642-54774-4{\_}3},
 year = {2014}
 }
 
-@phdthesis{Sumpf2012,
-author = {Sumpf, Tilman},
-file = {:mnt/data/Papers/ASL/ASL.bib:bib},
-keywords = {model-based reconstruction undersampling MRI T2 re},
-title = {{Model-based T 2 Relaxometry using Undersampled Magnetic Resonance Imaging}},
-school = {{Technical University of Braunschweig}},
-year = {2012},
-doi = {10.24355/dbbs.084-201305311128-0}
+@PhdThesis{Sumpf2012,
+  author = 	{Sumpf, Tilman},
+  title = 	{Model-based T2 Relaxometry using Undersampled Magnetic Resonance Imaging},
+  series = 	{Berichte aus dem Institut f{\"u}r Elektrische Messtechnik und Grundlagen der Elektrotechnik},
+  year = 	{2012},
+  month = 	{Dec},
+  day = 	{11},
+  publisher = 	{Mensch und Buch},
+  address = 	{Berlin},
+  volume = 	{43},
+  abstract = 	{T2 relaxometry refers to the quantitative determination of spin-spin relaxation times in magnetic resonance imaging (MRI). Particularly in clinical diagnostics, the method
+provides important information about tissue structures and respective pathologic alterations. Unfortunately, it also requires comparatively long measurement times which
+preclude widespread practical applications. To overcome such limitations, a so-called model-based reconstruction concept has recently been proposed. The method allows
+for the estimation of spin-density and T2 parameter maps from only a fraction of the usually required data. So far, promising results have been reported for a radial data
+acquisition scheme. However, due to technical reasons, radial imaging is only available on a very limited number of MRI systems.
+The present work deals with the realization and evaluation of different model-based T2 reconstruction methods that are applicable for the most widely available Cartesian
+(rectilinear) acquisition scheme. The initial implementation is based on the conventional assumption of a mono-exponential T2 signal decay. A suitable sampling scheme as well as an automatic scaling procedure are developed, which remove the necessity of manual parameter tuning. As demonstrated for human brain MRI data, the technique allows for a more than 5-fold acceleration of the underlying data acquisition. Furthermore, general limitations and specific error sources are identified and suitable simulation programs are developed for their detailed analysis. In addition to phase variations in image space, the simulations reveal truncation effects as a relevant cause of reconstruction artifacts. To reduce the latter, an alternative model formulation is developed and tested. For noisefree simulated data, the method yields an almost complete suppression of associated artifacts. Residual problems in the reconstruction of experimental MRI data point to the predominant influence of other errors in practice.
+The last part of this thesis focuses on the development of a refined T2 reconstruction technique which employs a signal model that considers contributions from stimulated
+echoes to the spin-echo signal. The method yields an increased accuracy of the estimated T2 relaxation times. In comparison, however, the mono-exponential model proves to
+be less sensitive to artifacts when the data acquisition is highly accelerated. This T2 relaxometry method is currently evaluated in a first clinical trial.},
+  doi = 	{10.24355/dbbs.084-201305311128-0},
+  url = 	{https://publikationsserver.tu-braunschweig.de/receive/dbbs_mods_00051567},
+  url = 	{http://www.digibib.tu-bs.de/?docid=00051567},
+  file = 	{:https://publikationsserver.tu-braunschweig.de/servlets/MCRFileNodeServlet/dbbs_derivate_00030132/Dissertation_Sumpf.pdf:PDF}
 }
 
 @article{Roeloffs2016,
