+++ .gitlab-ci.yml
@@ -1,33 +1,0 @@
-default:
-    image: registry.gitlab.tugraz.at/f23b736137140d66/mbpq
-
-build-job:
-    stage: build
-    script:
-        pip3 install -r requirements.txt
-        pip3 install -e .
-
-lint-job: 
-    stage: lint
-    script:
-        pylint -ry --output-format=parseable --exit-zero ./pyqmri > pylint.log
-    artifacts: pylint.log
-
-test-job:
-    stage: test
-    script:
-        pytest --junitxml results_unittests_LinOp.xml --cov=pyqmri test/unittests/test_LinearDataOperator.py
-        pytest --junitxml results_unittests_grad.xml --cov=pyqmri test/unittests/test_gradient.py
-        pytest --junitxml results_unittests_symgrad.xml --cov=pyqmri test/unittests/test_symmetrized_gradient.py
-    artifacts: /*.xml
-
-integration-job:
-    stage: integration-testing
-    script:
-        ipcluster start&
-        pytest --junitxml results_integrationtests_single_slice.xml --cov=pyqmri --integration-cover test/integrationtests/test_integration_test_single_slice.py
-        pytest --junitxml results_integrationtests_single_slice.xml --cov=pyqmri --integration-cover test/integrationtests/test_integration_test_multi_slice.py
-        ipcluster stop&
-    artifacts: /*.xml
-
-    
