+++ .gitlab-ci.yml
@@ -1,6 +1,14 @@
 default:
     image: registry.gitlab.tugraz.at/f23b736137140d66/mbpq
 
+variables:
+  PIP_CACHE_DIR: "$CI_PROJECT_DIR/.cache/pip"
+
+cache:
+  paths:
+    - .cache/pip
+    - venv/
+
 stages:
     - build
     - lint
