+++ .gitlab-ci.yml
@@ -19,7 +19,7 @@ build-job:
     stage: build
     script:
         - pip3 install -r requirements.txt
-        - pip3 install -e .
+        - pip3 install .
     tags:
         - fbmt129
 
