+++ .gitlab-ci.yml
@@ -35,6 +35,8 @@ lint-job:
 
 test-job:
     stage: test
+    before_script:
+        - pip install -e .
     script:
         - pytest --junitxml results_unittests_LinOp.xml --cov=pyqmri test/unittests/test_LinearDataOperator.py
         - pytest --junitxml results_unittests_grad.xml --cov=pyqmri test/unittests/test_gradient.py
@@ -47,6 +49,8 @@ test-job:
 
 integration-job:
     stage: integration-test
+    before_script:
+        - pip install -e .
     script:
         - ipcluster start&
         - pytest --junitxml results_integrationtests_single_slice.xml --cov=pyqmri --integration-cover test/integrationtests/test_integration_test_single_slice.py
