+++ .gitlab-ci.yml
@@ -15,6 +15,8 @@ cache:
 code_quality:
   artifacts:
     paths: [gl-code-quality-report.json]
+    tags:
+        - fbmt129
 
 stages:
     - build
