+++ .gitlab-ci.yml
@@ -13,8 +13,8 @@ cache:
     - venv/
 
 code_quality:
-  artifacts:
-    paths: [gl-code-quality-report.json]
+    artifacts:
+        paths: [gl-code-quality-report.json]
     tags:
         - fbmt129
 
