+++ .gitlab-ci.yml
@@ -1,7 +1,8 @@
 default:
     image: registry.gitlab.tugraz.at/f23b736137140d66/mbpq
-    tags:
-        - fbmt129
+tags:
+    - fbmt129
+    
 variables:
   PIP_CACHE_DIR: "$CI_PROJECT_DIR/.cache/pip"
 
