+++ .gitlab-ci.yml
@@ -1,7 +1,5 @@
 default:
     image: registry.gitlab.tugraz.at/f23b736137140d66/mbpq
-tags:
-    - fbmt129
     
 variables:
   PIP_CACHE_DIR: "$CI_PROJECT_DIR/.cache/pip"
@@ -21,6 +19,8 @@ build-job:
     script:
         - pip3 install -r requirements.txt
         - pip3 install .
+    tags:
+        - fbmt129
 
 test-job:
     stage: test
@@ -33,6 +33,8 @@ test-job:
     artifacts: 
         paths:
             - /*.xml
+    tags:
+        - fbmt129
 
 integration-job:
     stage: integration-test
@@ -46,5 +48,7 @@ integration-job:
     artifacts: 
         paths:
             - /*.xml
+    tags:
+        - fbmt129
 
     
