+++ .gitlab-ci.yml
@@ -9,14 +9,6 @@ stages:
 include:
   - template: Code-Quality.gitlab-ci.yml
 
-variables:
-  PIP_CACHE_DIR: "$CI_PROJECT_DIR/.cache/pip"
-
-cache:
-  paths:
-    - .cache/pip
-    - venv/
-
 code_quality:
     artifacts:
         paths: [gl-code-quality-report.json]
