+++ .gitlab-ci.yml
@@ -22,7 +22,8 @@ build-job:
 test-job:
     stage: test
     before_script:
-        - pip3 install -e .
+        - pip3 install -r requirements.txt --no-cache-dir
+        - pip3 install -e . --no-cache-dir
     script:
         - pytest --junitxml results_unittests_LinOp.xml --cov=pyqmri test/unittests/test_LinearDataOperator.py
         - pytest --junitxml results_unittests_grad.xml --cov=pyqmri test/unittests/test_gradient.py
@@ -34,7 +35,8 @@ test-job:
 integration-job:
     stage: integration-test
     before_script:
-        - pip3 install -e .
+        - pip3 install -r requirements.txt --no-cache-dir
+        - pip3 install -e . --no-cache-dir
     script:
         - ipcluster start&
         - pytest --junitxml results_integrationtests_single_slice.xml --cov=pyqmri --integration-cover test/integrationtests/test_integration_test_single_slice.py
