+++ .gitlab-ci.yml
@@ -29,11 +29,15 @@ test-job:
         - pip3 install -e . --no-cache-dir
     script:
         - pytest --junitxml results_unittests_LinOp.xml --cov=pyqmri test/unittests/test_LinearDataOperator.py
+        - coverage xml -o coverage_unittest_LinOp.xml
         - pytest --junitxml results_unittests_grad.xml --cov=pyqmri test/unittests/test_gradient.py
+        - coverage xml -o coverage_unittest_grad.xml
         - pytest --junitxml results_unittests_symgrad.xml --cov=pyqmri test/unittests/test_symmetrized_gradient.py
+        - coverage xml -o coverage_unittest_symgrad.xml
     artifacts: 
-        paths:
-            - /*.xml
+        reports:    
+            junit: result_unittests_*.xml
+            cobertura: coverage_unittest_*.xml
 
 integration-job:
     stage: integration-test
@@ -43,10 +47,13 @@ integration-job:
     script:
         - ipcluster start&
         - pytest --junitxml results_integrationtests_single_slice.xml --cov=pyqmri --integration-cover test/integrationtests/test_integration_test_single_slice.py
+        - coverage xml -o coverage_integrationtest_single_slice.xml
         - pytest --junitxml results_integrationtests_single_slice.xml --cov=pyqmri --integration-cover test/integrationtests/test_integration_test_multi_slice.py
+        - coverage xml -o coverage_integrationtest_multi_slice.xml
         - ipcluster stop&
     artifacts: 
-        paths:
-            - /*.xml
+        reports:    
+            junit: result_integrationtest_*.xml
+            cobertura: coverage_unittest_*.xml
 
     
