+++ .gitlab-ci.yml
@@ -24,7 +24,8 @@ lint-job:
     script:
         pylint -ry --output-format=parseable --exit-zero ./pyqmri > pylint.log
     artifacts:
-        paths: pylint.log
+        paths: 
+            - pylint.log
 
 test-job:
     stage: test
