+++ .gitlab-ci.yml
@@ -1,5 +1,5 @@
 default:
-    image: registry.gitlab.tugraz.at/f23b736137140d66/mbpq
+    image: registry.gitlab.tugraz.at/f23b736137140d66/pyqmri
 
 stages:
     - build
