+++ .gitlab-ci.yml
@@ -38,7 +38,7 @@ test-job:
         - pytest --junitxml results_unittests_grad.xml --cov=pyqmri test/unittests/test_gradient.py
 #        - coverage xml -o coverage_unittest_grad.xml
         - pytest --junitxml results_unittests_symgrad.xml --cov=pyqmri test/unittests/test_symmetrized_gradient.py
-        - coverage xml -o coverage_unittest.xml --include=results_unittests_'.*'.xml
+        - coverage xml -o coverage_unittest.xml --include=results_unittests_LinOp.xml,results_unittests_grad.xml,results_unittests_symgrad.xml
     artifacts: 
         reports:    
             junit: results_unittests_*.xml
@@ -54,7 +54,7 @@ integration-job:
         - pytest --junitxml results_integrationtests_single_slice.xml --cov=pyqmri --integration-cover test/integrationtests/test_integration_test_single_slice.py
 #        - coverage xml -o coverage_integrationtests_single_slice.xml
         - pytest --junitxml results_integrationtests_multi_slice.xml --cov=pyqmri --integration-cover test/integrationtests/test_integration_test_multi_slice.py
-        - coverage xml -o coverage_integrationtest.xml --include=results_integrationtests_'.*'.xml
+        - coverage xml -o coverage_integrationtest.xml --include=results_integrationtests_single_slice.xml,results_integrationtests_multi_slice.xml
         - ipcluster stop&
     artifacts: 
         reports:    
