### 2020-11-24 #2727 @MaierOli2010
@agahkarakuzu I've moved the CI from Jenkins to the University GitLab site. This should provide similar functionality but without the security vulnerabilities of exposing my own workstation to the public. The GitLab repo is synced with GitHub and all pushed to GitLab will be mirrored to GitHub as well. I've updated the `Contributing` section to point to GitLab for Code development.
