+++ README.rst
@@ -45,7 +45,7 @@ entry in the .h5 to force a recomputation of the receive coil sensitivities as t
 
 Contributing
 ------------
-Development and code contributaions should be done at our GitLab_ site to facilitate the CI integration and GPU availability there.
+Development and code contributions should be done at our GitLab_ site to facilitate the CI integration and GPU availability there.
 If you want to contribute please make sure that all tests pass and adhere to our `Code of Conduct`_. 
 Prior to running the tests it is necessary to start an ipcluster. 
 An exemplary workflow would be:
