### 2020-11-24 30871de
Rm text from badges
