### 2020-11-24 #2727 @agahkarakuzu
@MaierOli2010 I made a small PR to fix a typo. Other than that, looks like all my comments have been resolved, thank you so much for addressing them and making this amazing tool publicly available!

I checked all the remaining boxes, all good on my end.
