### 2020-11-26 90c96b9
Modified plotting to account for mutliple PLD
