+++ pyqmri/models/ASL.py
@@ -107,7 +107,7 @@ class Model(BaseModel):
         self.constraints.append(
             constraints(0,
                         200,
-                        True))
+                        False))
         self.constraints.append(
             constraints(0.01/60,
                         5/60,
