### 2020-11-26 e66dfd9
Updated initial guess. Working with all data sets currently
