+++ test/unittests/test_gradient.py
@@ -35,8 +35,8 @@ def setupPar(par):
     par["unknowns_TGV"] = 2
     par["unknowns_H1"] = 0
     par["unknowns"] = 2
-    par["dz"] = 3
-    par["weights"] = np.array([1, 0.1])
+    par["dz"] = 1
+    par["weights"] = np.array([1, 1])
 
 
 class GradientTest(unittest.TestCase):
@@ -81,7 +81,6 @@ class GradientTest(unittest.TestCase):
         self.divin = self.divin.astype(DTYPE)
         self.dz = par["dz"]
         self.queue = par["queue"][0]
-        self.weights = par["weights"]
 
     def test_grad_outofplace(self):
         gradx = np.zeros_like(self.gradin)
@@ -95,14 +94,12 @@ class GradientTest(unittest.TestCase):
         grad = np.stack((gradx,
                          grady,
                          gradz), axis=-1)
-        for j in range(grad.shape[0]):
-            grad[j] *= self.weights[j]
 
         inp = clarray.to_device(self.queue, self.gradin)
         outp = self.grad.fwdoop(inp)
         outp = outp.get()
 
-        np.testing.assert_allclose(outp[..., :-1], grad, rtol=1e-15)
+        np.testing.assert_allclose(outp[..., :-1], grad, rtol=0)
 
     def test_grad_inplace(self):
         gradx = np.zeros_like(self.gradin)
@@ -116,16 +113,13 @@ class GradientTest(unittest.TestCase):
         grad = np.stack((gradx,
                          grady,
                          gradz), axis=-1)
-        
-        for j in range(grad.shape[0]):
-            grad[j] *= self.weights[j]
-            
+
         inp = clarray.to_device(self.queue, self.gradin)
         outp = clarray.to_device(self.queue, self.divin)
         self.grad.fwd(outp, inp)
         outp = outp.get()
 
-        np.testing.assert_allclose(outp[..., :-1], grad, rtol=1e-15)
+        np.testing.assert_allclose(outp[..., :-1], grad, rtol=0)
 
     def test_adj_outofplace(self):
         inpgrad = clarray.to_device(self.queue, self.gradin)
@@ -209,7 +203,6 @@ class GradientStreamedTest(unittest.TestCase):
         self.gradin = self.gradin.astype(DTYPE)
         self.divin = self.divin.astype(DTYPE)
         self.dz = par["dz"]
-        self.weights = par["weights"]
 
     def test_grad_outofplace(self):
         gradx = np.zeros_like(self.gradin)
@@ -223,13 +216,10 @@ class GradientStreamedTest(unittest.TestCase):
         grad = np.stack((gradx,
                          grady,
                          gradz), axis=-1)
-        
-        for j in range(grad.shape[1]):
-            grad[:,j] *= self.weights[j]
 
         outp = self.grad.fwdoop([[self.gradin]])
 
-        np.testing.assert_allclose(outp[..., :-1], grad, rtol=1e-15)
+        np.testing.assert_allclose(outp[..., :-1], grad, rtol=0)
 
     def test_grad_inplace(self):
         gradx = np.zeros_like(self.gradin)
@@ -243,15 +233,12 @@ class GradientStreamedTest(unittest.TestCase):
         grad = np.stack((gradx,
                          grady,
                          gradz), axis=-1)
-        
-        for j in range(grad.shape[1]):
-            grad[:,j] *= self.weights[j]
 
         outp = np.zeros_like(self.divin)
 
         self.grad.fwd([outp], [[self.gradin]])
 
-        np.testing.assert_allclose(outp[..., :-1], grad, rtol=1e-15)
+        np.testing.assert_allclose(outp[..., :-1], grad, rtol=0)
 
     def test_adj_outofplace(self):
 
