+++ .gitlab-ci.yml
@@ -1,0 +1,76 @@
+default:
+    image: registry.gitlab.tugraz.at/f23b736137140d66/pyqmri
+
+stages:
+    - build
+    - linting
+    - test
+    - integration-test
+
+# include:
+#  - template: Code-Quality.gitlab-ci.yml
+#code_quality:
+#    artifacts:
+#        paths: [gl-code-quality-report.json]
+
+build-job:
+    stage: build
+    script:
+        - pip3 install -r requirements.txt
+        - pip3 install .
+
+lint-job:
+    stage: linting
+    script:
+        pylint -ry --output-format=parseable --exit-zero ./pyqmri > pylint.log
+    artifacts:
+        paths: 
+            - pylint.log
+
+test-job:
+    stage: test
+    before_script:
+        - pip3 install -r requirements.txt --no-cache-dir
+        - pip3 install -e . --no-cache-dir
+    script:
+        - pytest --junitxml results_unittests_LinOp.xml --cov=pyqmri test/unittests/test_LinearDataOperator.py
+        - coverage xml -o coverage_unittests_LinOp.xml
+        - pytest --junitxml results_unittests_grad.xml --cov=pyqmri test/unittests/test_gradient.py
+        - coverage xml -o coverage_unittests_grad.xml
+        - pytest --junitxml results_unittests_symgrad.xml --cov=pyqmri test/unittests/test_symmetrized_gradient.py
+        - coverage xml -o coverage_unittests_symgrad.xml
+    artifacts: 
+        reports:    
+            junit: results_unittests_*.xml
+            cobertura: coverage_unittests_*.xml
+
+integration-job:
+    stage: integration-test
+    before_script:
+        - pip3 install -r requirements.txt --no-cache-dir
+        - pip3 install -e . --no-cache-dir
+    script:
+        - ipcluster start&
+        - pytest --junitxml results_integrationtests_single_slice.xml --cov=pyqmri --integration-cover test/integrationtests/test_integration_test_single_slice.py
+        - coverage xml -o coverage_integrationtests_single_slice.xml
+        - pytest --junitxml results_integrationtests_multi_slice.xml --cov=pyqmri --integration-cover test/integrationtests/test_integration_test_multi_slice.py
+        - coverage xml -o coverage_integrationtests_multi_slice.xml
+        - ipcluster stop&
+    artifacts: 
+        reports:    
+            junit: results_integrationtests_*.xml
+            cobertura: coverage_integrationtests_*.xml
+
+#pages:
+#  stage: deploy
+#  dependencies:
+#    - test-job
+#    - integration-job
+#  script:
+#    - mv coverage*.xml public/
+#  artifacts:
+#    paths:
+#      - public
+#    expire_in: 30 days
+#  only:
+#    - master    
