### 2020-11-26 e7e19d2
calc image metrics; data export and import functionality; attempt streamed version of soft sense operator --> not working (unit test ok)
