+++ main.py
@@ -1,4 +1,6 @@
 import argparse
+import sys
+
 import pyopencl.array as cla
 import pyqmri.operator as pyqmirop
 import pyqmri.solver as pyqmrisl
@@ -36,13 +38,14 @@ def _setup_par(par, myargs, ksp_data):
     par["dz"] = 1
 
     par["fft_dim"] = (0, 1, 2)  # 3D fft
-    par["mask"] = np.ones(ksp_data[0, 0, ...].shape, dtype=DTYPE_real)
+    #par["mask"] = np.ones(ksp_data[0, 0, ...].shape, dtype=DTYPE_real)
+    par["mask"] = np.require(np.ones((par["dimY"], par["dimX"]), dtype=DTYPE_real), requirements='C')
 
     par["overlap"] = 0
     par["par_slices"] = par["NSlice"]
 
     if myargs.streamed:
-        par["par_slices"] = par["NSlice"] / par["num_dev"]
+        par["par_slices"] = 4
         par["overlap"] = 1
 
 
@@ -51,7 +54,7 @@ def _setup_irgn_par(irgn_par, myargs):
     irgn_par["accelerated"] = False
     irgn_par["tol"] = 1e-8
     irgn_par["stag"] = 1000
-    irgn_par["sigma"] = 1
+    irgn_par["sigma"] = np.float32(1 / np.sqrt(12))
     irgn_par["lambd"] = myargs.lamda
     irgn_par["alpha0"] = np.sqrt(2)  # 2D --> np.sqrt(2), 3D --> np.sqrt(3)
     irgn_par["alpha1"] = 1   # 1
@@ -169,7 +172,7 @@ def _pda_soft_sense_solver(myargs, par, ksp, cmaps, imgs, imgs_us, reg_type=''):
     inp = inp.astype(DTYPE)
 
     # if reg_type != '':
-    irgn_par["sigma"] = np.float32(1 / np.sqrt(12))
+    # irgn_par["sigma"] = np.float32(1 / np.sqrt(12))
 
     fval = calculate_cost(myargs, irgn_par, inp, np.zeros(inp.shape+(3,)), ksp, cmaps, reg_type)
 
@@ -220,14 +223,28 @@ def _2d_recon(imgs, ksp_data, cmaps, par, myargs):
     imgs = imgs[:, 21:21+NSlice, ...]
     #imgs = np.expand_dims(imgs, axis=0)
 
+    if myargs.streamed:
+        data_trans_axes = (2, 0, 1, 3, 4)
+        ksp_data2d = np.transpose(ksp_data2d, data_trans_axes)
+        cmaps = np.transpose(cmaps, data_trans_axes)
+
+        #ksp_data2d = np.moveaxis(ksp_data2d, [0, 1, 2], [1, 2, 0])
+        #cmaps = np.moveaxis(cmaps, [0, 1, 2], [1, 2, 0])
+
     out_undersampled = soft_sense_recon_cl(myargs, par, ksp_data2d, cmaps)
-    img_montage(sum_of_squares(out_undersampled), '2D reconstruction of undersampled data with adjoint operator')
+
+    if myargs.streamed:
+        out_undersampled = np.moveaxis(out_undersampled, [0, 1], [1, 0])
+        save_imgs(sum_of_squares(out_undersampled), myargs, 'test_streamed')
+        sys.exit()
+
+    # img_montage(sum_of_squares(out_undersampled), '2D reconstruction of undersampled data with adjoint operator')
 
     out_pd = _pda_soft_sense_solver(myargs, par, ksp_data2d, cmaps, imgs, out_undersampled, reg_type=myargs.reg_type)
 
     x_orig = normalize_imgs(sum_of_squares(imgs))
     x_ssense = normalize_imgs(sum_of_squares(out_pd))
[…]
