### 2020-11-27 #2727 @MaierOli2010
Had to update affiliations. Should be all set now from my side. I thought about increasing the version number to 1.0, i.e., first release if the paper is accepted. Should I do that before or prior to the published paper?
