### 2020-11-27 44fd7d4
Streamed version of SoftSenseOperator implemented; some minor refactoring
