+++ pyqmri/models/FFCPaperMultiAlpha.py
@@ -41,7 +41,7 @@ class Model(BaseModel):
         self.constraints.append(
             constraints(0,
                         2,
-                        True))
+                        False))
         self.constraints.append(
             constraints(1,
                         2000,
@@ -55,8 +55,15 @@ class Model(BaseModel):
             self.constraints.append(
                 constraints(0,
                             2,
-                            True))
-
+                            False))
+        self._ind1 = 0
+        self._ind2 = 0
+        self._labels = []
+        for j in range(len(self.b)):
+            self._labels.append(
+                "Field "+str(np.round(self.b[j]*1e3, 2))+" mT")
+            
+            
     def rescale(self, x):
         tmp_x = np.copy(x)
         tmp_x[0] *= self.uk_scale[0]
@@ -198,137 +205,166 @@ class Model(BaseModel):
         grad[~np.isfinite(grad)] = 1e-20
 
         return grad
+    def plot_unknowns(self, x, dim_2D=False):
+        unknowns = self.rescale(x)
+        tmp_x = unknowns["data"]
 
-    # def plot_unknowns(self, x, dim_2D=False):
-    #         unknowns = self.rescale(x)
-    #         tmp_x = unknowns["data"]
-    #         images = np.abs(self._execute_forward_3D(x) / self.dscale)[
-    #             :self.t.shape[1]]
-    #         # tmp_x = (self.rescale(x))
-    #         tmp_x[0] = np.abs(tmp_x[0])/self.dscale
-    #         # tmp_x[1] /= self.dscale
-    #         tmp_x[1] = np.abs(tmp_x[1])
-    #         tmp_x = np.real(tmp_x)
+        images = np.abs(self._execute_forward_3D(x) / self.dscale)
+        images = np.reshape(images, self.t.shape+images.shape[-3:])
 
-    #         ind1 = 26
-    #         ind2 = 26
+        tmp_x[0] = np.abs(tmp_x[0])/self.dscale
+        tmp_x[1] = np.abs(tmp_x[1])
+        tmp_x[-self.numT1Scale:] = np.abs(tmp_x[-self.numT1Scale:])
+        tmp_x = np.real(tmp_x)
 
-    #         if dim_2D:
-    #             pass
-    #         else:
-    #             self.ax = []
-    #             if not self._figure:
-    #                 plot_dim = int(np.ceil(np.sqrt(len(self.uk_scale))))
-    #                 plt.ion()
-    #                 self._figure = plt.figure(figsize=(12, 6))
-    #                 self._figure.subplots_adjust(hspace=0.3, wspace=0)
-    #                 wd_ratio = np.tile([1, 1 / 20, 1 / (5)], plot_dim)
-    #                 self.gs = gridspec.GridSpec(
-    #                     plot_dim+1, 3*plot_dim,
-    #                     width_ratios=wd_ratio, hspace=0.3, wspace=0)
-    #                 self._figure.tight_layout()
-    #                 self._figure.patch.set_facecolor(plt.cm.viridis.colors[0])
-    #                 for grid in self.gs:
-    #                     self.ax.append(plt.subplot(grid))
-    #                     self.ax[-1].axis('off')
-    #                 self._plot = []
-    #                 for j in range(len(self.uk_scale)):
-    #                     self._plot.append(
-    #                         self.ax[3 * j].imshow(
-    #                             tmp_x[j, int(self.NSlice / 2), ...]))
-    #                     self.ax[3 *
-    #                             j].set_title('UK : ' +
-    #                                          str(j), color='white')
-    #                     self.ax[3 * j + 1].axis('on')
-    #                     cbar = self._figure.colorbar(
-    #                         self._plot[j], cax=self.ax[3 * j + 1])
[…]
