### 2020-12-01 6cb958c
Updated kernels to allow 2D reg
