+++ pyqmri/models/FFCPaperPhase.py
@@ -23,15 +23,15 @@ class Model(BaseModel):
             self.b = self.b[None]
             self.t = self.t[None]
 
-        self.numT1Scale = len(self.b)-1
+        self.numT1Scale = len(self.b)
 
-        par["unknowns_TGV"] = 2 + len(self.b) + len(self.b) - 1
+        par["unknowns_TGV"] = 1 + len(self.b) + len(self.b)
         par["unknowns_H1"] = 0
         par["unknowns"] = par["unknowns_TGV"]+par["unknowns_H1"]
 
         self.unknowns = par["unknowns"]
         
-        t1min = np.min(self.t)/2
+        t1min = 0#np.min(self.t)/2
 
         for j in range(par["unknowns"]):
             self.uk_scale.append(1)
@@ -40,24 +40,16 @@ class Model(BaseModel):
             constraints(0,
                         1000,
                         True))
-        self.constraints.append(
-            constraints(0,
-                        1,
-                        False))
-        self.constraints.append(
-            constraints(t1min,
-                        2000,
-                        True))
-        for j in range(self.numT1Scale):
-            self.constraints.append(
-                constraints(t1min,
-                            2000,
-                            True))
         for j in range(self.numT1Scale):
             self.constraints.append(
                 constraints(0,
                             1,
                             False))
+        for j in range(self.numT1Scale):
+            self.constraints.append(
+                constraints(t1min,
+                            10000,
+                            True))
         self._ind1 = 0
         self._ind2 = 0
         self._labels = []
@@ -68,15 +60,13 @@ class Model(BaseModel):
     def rescale(self, x):
         tmp_x = np.copy(x)
         tmp_x[0] *= self.uk_scale[0]
-        tmp_x[1] *= self.uk_scale[1]
-        tmp_x[2] *= self.uk_scale[2]
-        ukname = ["C", "alpha", "T1_1"]
+        ukname = ["C", "alpha"]
         for j in range(self.numT1Scale):
-            tmp_x[3+j] *= self.uk_scale[3+j]
-            ukname.append("T1_"+str(2+j))
+            tmp_x[1+j] *= self.uk_scale[1+j]
+            ukname.append("alpha_"+str(1+j))
         for j in range(self.numT1Scale):
-            tmp_x[3+self.numT1Scale+j] *= self.uk_scale[3+self.numT1Scale+j]
-            ukname.append("al_"+str(2+j))
+            tmp_x[-self.numT1Scale+j] *= self.uk_scale[-self.numT1Scale+j]
+            ukname.append("T1_"+str(1+j))
         const = []
         for constrained in self.constraints:
             const.append(constrained.real)
@@ -89,22 +79,15 @@ class Model(BaseModel):
             (self.NScan, self.NSlice, self.dimY, self.dimX),
             dtype=self._DTYPE)
         t = self.t[0][:, None, None, None]
-        S[:len(t)] = (
-            x[0] * self.uk_scale[0]*x[1]*self.uk_scale[1]
-            * (-self.b0 *
-               np.exp(-t / (x[2] * self.uk_scale[2]))
-               + (1 - np.exp(-t / (x[2] * self.uk_scale[2])))
-               * self.b[0])
-            )
         for j in range(self.numT1Scale):
-            offset = len(self.t[j+1])
[…]
