+++ pyqmri/models/FFCPaperPhaseR1.py
@@ -43,12 +43,12 @@ class Model(BaseModel):
         for j in range(self.numT1Scale):
             self.constraints.append(
                 constraints(0,
-                            10,
+                            1,
                             False))
         for j in range(self.numT1Scale):
             self.constraints.append(
-                constraints(t1min,
-                            600/(j+1),
+                constraints(1/(800/(j+1)),
+                            1/t1min,
                             True))
         self._ind1 = 0
         self._ind2 = 0
@@ -66,6 +66,7 @@ class Model(BaseModel):
             ukname.append("alpha_"+str(1+j))
         for j in range(self.numT1Scale):
             tmp_x[-self.numT1Scale+j] *= self.uk_scale[-self.numT1Scale+j]
+            tmp_x[-self.numT1Scale+j] = 1/tmp_x[-self.numT1Scale+j]
             ukname.append("T1_"+str(1+j))
         const = []
         for constrained in self.constraints:
@@ -85,8 +86,8 @@ class Model(BaseModel):
             S[offset*(j):offset*(j+1)] = (
                 x[0] * self.uk_scale[0]*x[1+j]*self.uk_scale[1+j]
                 * (-self.b0 *
-                   np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
-                   + (1 - np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
+                   np.exp(-t * (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
+                   + (1 - np.exp(-t * (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
                    * self.b[j])
                 )
         S[~np.isfinite(S)] = 1e-20
@@ -113,8 +114,8 @@ class Model(BaseModel):
             grad[offset*(j):offset*(j+1)] = (
                 self.uk_scale[0]*x[1+j]*self.uk_scale[1+j]
                 * (-self.b0 *
-                   np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
-                   + (1 - np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
+                   np.exp(-t * (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
+                   + (1 - np.exp(-t * (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
                    * self.b[j])
                 )
         grad[~np.isfinite(grad)] = 1e-20
@@ -132,8 +133,8 @@ class Model(BaseModel):
             grad[j, offset*(j):offset*(j+1)] = (
                 x[0] * self.uk_scale[0]*self.uk_scale[1+j]
                 * (-self.b0 *
-                   np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
-                   + (1 - np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
+                   np.exp(-t * (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
+                   + (1 - np.exp(-t * (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
                    * self.b[j])
                 )
         grad[~np.isfinite(grad)] = 1e-20
@@ -151,11 +152,11 @@ class Model(BaseModel):
             grad[j, (j)*offset:(j+1)*offset] = (
[…]
