### 2020-12-03 b062adf
fixing beta of linesearch
