### 2020-12-08 #2727 @Kevin-Mattheus-Moerman
At this point could you do the following:

- [ ] Archive a copy of the software on [ZENODO](https://zenodo.org/) (some find these steps useful: https://guides.github.com/activities/citable-code/)? Please ensure that the metadata, in particular the author information, names, and order match the information for the JOSS paper. Can you also make sure the title matches the JOSS paper? You may need to manually edit this information. Once you've archived a copy please report back here with the archive DOI. Thanks.

- [ ] Can you confirm what the latest version tag is for this work? Is it still at `v0.3.1`?

- [ ] Since your paper is about to be accepted I recommend you read through it once more yourself and make any final edits if needed.
