+++ pyqmri/models/FFCPaperPhaseAlpha.py
@@ -25,7 +25,7 @@ class Model(BaseModel):
 
         self.numT1Scale = len(self.b)
 
-        par["unknowns_TGV"] = 1 + len(self.b) + len(self.b)
+        par["unknowns_TGV"] = len(self.b) + len(self.b) + len(self.b)
         par["unknowns_H1"] = 0
         par["unknowns"] = par["unknowns_TGV"]+par["unknowns_H1"]
 
@@ -35,16 +35,17 @@ class Model(BaseModel):
 
         for j in range(par["unknowns"]):
             self.uk_scale.append(1)
-
-        self.constraints.append(
-            constraints(1,
-                        1,
-                        True))
+            
         for j in range(self.numT1Scale):
             self.constraints.append(
                 constraints(0,
-                            100,
+                            1000,
                             False))
+        for j in range(self.numT1Scale):
+            self.constraints.append(
+                constraints(0,
+                            5,
+                            True))
         for j in range(self.numT1Scale):
             self.constraints.append(
                 constraints(t1min,
@@ -59,10 +60,12 @@ class Model(BaseModel):
 
     def rescale(self, x):
         tmp_x = np.copy(x)
-        tmp_x[0] *= self.uk_scale[0]
-        ukname = ["C"]
+        ukname = []
         for j in range(self.numT1Scale):
-            tmp_x[1+j] *= self.uk_scale[1+j]
+            tmp_x[j] *= self.uk_scale[j]
+            ukname.append("C_"+str(1+j))
+        for j in range(self.numT1Scale):
+            tmp_x[self.numT1Scale+j] *= self.uk_scale[self.numT1Scale+j]
             ukname.append("alpha_"+str(1+j))
         for j in range(self.numT1Scale):
             tmp_x[-self.numT1Scale+j] *= self.uk_scale[-self.numT1Scale+j]
@@ -83,8 +86,8 @@ class Model(BaseModel):
             offset = len(self.t[j])
             t = self.t[j][:, None, None, None]
             S[offset*(j):offset*(j+1)] = (
-                x[0] * self.uk_scale[0]*x[1+j]*self.uk_scale[1+j]
-                * (-self.b0 *
+                x[j] * self.uk_scale[j]
+                * (-self.b0 * x[self.numT1Scale+j]*self.uk_scale[self.numT1Scale+j] *
                    np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
                    + (1 - np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
                    * self.b[j])
@@ -99,20 +102,20 @@ class Model(BaseModel):
         gradAlpha = self._gradAlpha(x)
         gradT1 = self._gradT1(x)
 
-        grad = np.concatenate((gradC[None], gradAlpha, gradT1), axis=0)
+        grad = np.concatenate((gradC, gradAlpha, gradT1), axis=0)
         return grad
 
     def _gradC(self, x):
         grad = np.zeros(
-            (self.NScan, self.NSlice, self.dimY, self.dimX),
+            (self.numT1Scale, self.NScan, self.NSlice, self.dimY, self.dimX),
             dtype=self._DTYPE)
         t = self.t[0][:, None, None, None]
         for j in range(self.numT1Scale):
             offset = len(self.t[j])
             t = self.t[j][:, None, None, None]
-            grad[offset*(j):offset*(j+1)] = (
-                self.uk_scale[0]*x[1+j]*self.uk_scale[1+j]
-                * (-self.b0 *
[…]
