+++ pyqmri/models/FFCPaperPhaseM0.py
@@ -25,7 +25,7 @@ class Model(BaseModel):
 
         self.numT1Scale = len(self.b)
 
-        par["unknowns_TGV"] = 1 + len(self.b) + len(self.b)
+        par["unknowns_TGV"] = len(self.b) + len(self.b) + len(self.b)
         par["unknowns_H1"] = 0
         par["unknowns"] = par["unknowns_TGV"]+par["unknowns_H1"]
 
@@ -35,20 +35,21 @@ class Model(BaseModel):
 
         for j in range(par["unknowns"]):
             self.uk_scale.append(1)
-
-        self.constraints.append(
-            constraints(1,
-                        1,
-                        True))
+            
         for j in range(self.numT1Scale):
             self.constraints.append(
                 constraints(0,
-                            100,
+                            1000,
+                            False))
+        for j in range(self.numT1Scale):
+            self.constraints.append(
+                constraints(0,
+                            1000,
                             False))
         for j in range(self.numT1Scale):
             self.constraints.append(
                 constraints(t1min,
-                            2000/(j+1),
+                            500/(1),
                             True))
         self._ind1 = 0
         self._ind2 = 0
@@ -59,10 +60,12 @@ class Model(BaseModel):
 
     def rescale(self, x):
         tmp_x = np.copy(x)
-        tmp_x[0] *= self.uk_scale[0]
-        ukname = ["C"]
+        ukname = []
         for j in range(self.numT1Scale):
-            tmp_x[1+j] *= self.uk_scale[1+j]
+            tmp_x[j] *= self.uk_scale[j]
+            ukname.append("C_"+str(1+j))
+        for j in range(self.numT1Scale):
+            tmp_x[self.numT1Scale+j] *= self.uk_scale[self.numT1Scale+j]
             ukname.append("alpha_"+str(1+j))
         for j in range(self.numT1Scale):
             tmp_x[-self.numT1Scale+j] *= self.uk_scale[-self.numT1Scale+j]
@@ -83,11 +86,9 @@ class Model(BaseModel):
             offset = len(self.t[j])
             t = self.t[j][:, None, None, None]
             S[offset*(j):offset*(j+1)] = (
-                x[0] * self.uk_scale[0]*x[1+j]*self.uk_scale[1+j]
-                * (-self.b0 *
-                   np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
-                   + (1 - np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
-                   * self.b[j])
+                - x[j] * self.uk_scale[j] * np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
+                + (1 - np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
+                * x[self.numT1Scale+j]*self.uk_scale[self.numT1Scale+j]
                 )
         S[~np.isfinite(S)] = 1e-20
         S = np.array(S, dtype=self._DTYPE)
@@ -99,23 +100,21 @@ class Model(BaseModel):
         gradAlpha = self._gradAlpha(x)
         gradT1 = self._gradT1(x)
 
-        grad = np.concatenate((gradC[None], gradAlpha, gradT1), axis=0)
+        grad = np.concatenate((gradC, gradAlpha, gradT1), axis=0)
         return grad
 
     def _gradC(self, x):
         grad = np.zeros(
-            (self.NScan, self.NSlice, self.dimY, self.dimX),
[…]
