+++ pyqmri/models/FFCPaperPhaseM0.py
@@ -24,10 +24,12 @@ class Model(BaseModel):
             self.t = self.t[None]
 
         self.numT1Scale = len(self.b)
+        self.numC = 1
 
-        par["unknowns_TGV"] = len(self.b) + len(self.b) + len(self.b)
+        par["unknowns_TGV"] = self.numC + len(self.b) + len(self.b)
         par["unknowns_H1"] = 0
         par["unknowns"] = par["unknowns_TGV"]+par["unknowns_H1"]
+
 
         self.unknowns = par["unknowns"]
         
@@ -36,7 +38,7 @@ class Model(BaseModel):
         for j in range(par["unknowns"]):
             self.uk_scale.append(1)
             
-        for j in range(self.numT1Scale):
+        for j in range(self.numC):
             self.constraints.append(
                 constraints(0,
                             1000,
@@ -49,7 +51,7 @@ class Model(BaseModel):
         for j in range(self.numT1Scale):
             self.constraints.append(
                 constraints(t1min,
-                            500/(1),
+                            1000/(1),
                             True))
         self._ind1 = 0
         self._ind2 = 0
@@ -61,11 +63,11 @@ class Model(BaseModel):
     def rescale(self, x):
         tmp_x = np.copy(x)
         ukname = []
-        for j in range(self.numT1Scale):
+        for j in range(self.numC):
             tmp_x[j] *= self.uk_scale[j]
             ukname.append("C_"+str(1+j))
         for j in range(self.numT1Scale):
-            tmp_x[self.numT1Scale+j] *= self.uk_scale[self.numT1Scale+j]
+            tmp_x[self.numC+j] *= self.uk_scale[self.numC+j]
             ukname.append("alpha_"+str(1+j))
         for j in range(self.numT1Scale):
             tmp_x[-self.numT1Scale+j] *= self.uk_scale[-self.numT1Scale+j]
@@ -81,14 +83,13 @@ class Model(BaseModel):
         S = np.zeros(
             (self.NScan, self.NSlice, self.dimY, self.dimX),
             dtype=self._DTYPE)
-        t = self.t[0][:, None, None, None]
         for j in range(self.numT1Scale):
             offset = len(self.t[j])
             t = self.t[j][:, None, None, None]
             S[offset*(j):offset*(j+1)] = (
-                - x[j] * self.uk_scale[j] * np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
+                - self.b0*x[np.mod(j,self.numC)] * self.uk_scale[np.mod(j,self.numC)] * np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
                 + (1 - np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
-                * x[self.numT1Scale+j]*self.uk_scale[self.numT1Scale+j]
+                * x[self.numC+j]*self.uk_scale[self.numC+j]*self.b[j]
                 )
         S[~np.isfinite(S)] = 1e-20
         S = np.array(S, dtype=self._DTYPE)
@@ -105,14 +106,13 @@ class Model(BaseModel):
 
     def _gradC(self, x):
         grad = np.zeros(
-            (self.numT1Scale, self.NScan, self.NSlice, self.dimY, self.dimX),
+            (self.numC, self.NScan, self.NSlice, self.dimY, self.dimX),
             dtype=self._DTYPE)
-        t = self.t[0][:, None, None, None]
         for j in range(self.numT1Scale):
             offset = len(self.t[j])
             t = self.t[j][:, None, None, None]
[…]
