+++ pyqmri/models/FFCPaperPhaseAlpha.py
@@ -24,10 +24,13 @@ class Model(BaseModel):
             self.t = self.t[None]
 
         self.numT1Scale = len(self.b)
+        self.numC = len(self.b)
+        self.numAlpha = len(self.b)
 
-        par["unknowns_TGV"] = len(self.b) + len(self.b) + len(self.b)
+        par["unknowns_TGV"] = self.numC + self.numAlpha + len(self.b)
         par["unknowns_H1"] = 0
         par["unknowns"] = par["unknowns_TGV"]+par["unknowns_H1"]
+
 
         self.unknowns = par["unknowns"]
         
@@ -36,15 +39,15 @@ class Model(BaseModel):
         for j in range(par["unknowns"]):
             self.uk_scale.append(1)
             
-        for j in range(self.numT1Scale):
+        for j in range(self.numC):
             self.constraints.append(
                 constraints(0,
                             1000,
                             False))
-        for j in range(self.numT1Scale):
+        for j in range(self.numAlpha):
             self.constraints.append(
-                constraints(0,
-                            5,
+                constraints(1,
+                            1,
                             True))
         for j in range(self.numT1Scale):
             self.constraints.append(
@@ -61,11 +64,11 @@ class Model(BaseModel):
     def rescale(self, x):
         tmp_x = np.copy(x)
         ukname = []
-        for j in range(self.numT1Scale):
+        for j in range(self.numC):
             tmp_x[j] *= self.uk_scale[j]
             ukname.append("C_"+str(1+j))
-        for j in range(self.numT1Scale):
-            tmp_x[self.numT1Scale+j] *= self.uk_scale[self.numT1Scale+j]
+        for j in range(self.numAlpha):
+            tmp_x[self.numC+j] *= self.uk_scale[self.numC+j]
             ukname.append("alpha_"+str(1+j))
         for j in range(self.numT1Scale):
             tmp_x[-self.numT1Scale+j] *= self.uk_scale[-self.numT1Scale+j]
@@ -86,8 +89,8 @@ class Model(BaseModel):
             offset = len(self.t[j])
             t = self.t[j][:, None, None, None]
             S[offset*(j):offset*(j+1)] = (
-                x[j] * self.uk_scale[j]
-                * (-self.b0 * x[self.numT1Scale+j]*self.uk_scale[self.numT1Scale+j] *
+                x[np.mod(j,self.numC)] * self.uk_scale[np.mod(j,self.numC)]
+                * (-self.b0 * x[self.numC+np.mod(j,self.numAlpha)]*self.uk_scale[self.numC+np.mod(j,self.numAlpha)] *
                    np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
                    + (1 - np.exp(-t / (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
                    * self.b[j])
@@ -107,15 +110,15 @@ class Model(BaseModel):
 
     def _gradC(self, x):
         grad = np.zeros(
-            (self.numT1Scale, self.NScan, self.NSlice, self.dimY, self.dimX),
+            (self.numC, self.NScan, self.NSlice, self.dimY, self.dimX),
             dtype=self._DTYPE)
         t = self.t[0][:, None, None, None]
         for j in range(self.numT1Scale):
             offset = len(self.t[j])
             t = self.t[j][:, None, None, None]
-            grad[j, offset*(j):offset*(j+1)] = (
-                self.uk_scale[j]
[…]
