### 2020-12-09 #2727 @MaierOli2010
@Kevin-Mattheus-Moerman
I've archived the software. The associated DOI is: 10.5281/zenodo.4313301
Further, I decided to push the version to "first release", i.e. v1.0.0 as the acceptance of the paper marks the first release in my opinion. This is also the archived version tag.

There were some minor edits to affiliations and I've shortened a sentence in the introduction as its meaning was the same as of the one beforehand.

Finally, I would like to thank all of you for reviewing our work :)
