### 2020-12-10 3fd1ffe
Changed updates in PD algo by implicit calculations; Removed fftshift through multiplying by checker from Soft Sense operator
