+++ pyqmri/solver.py
@@ -2496,6 +2496,18 @@ class PDSoftSenseBaseSolver:
                         linops,
                         coils,
                         model)
+            else:
+                if not SMS:
+                    pdop = PDSoftSenseSolverStreamedTGV(
+                        par,
+                        irgn_par,
+                        queue,
+                        np.float32(1 / np.sqrt(8)),
+                        init_fval,
+                        prg,
+                        linops,
+                        coils,
+                        model)
         else:
             raise NotImplementedError
         return pdop
@@ -2559,11 +2571,6 @@ class PDSoftSenseBaseSolver:
                 sigma=sigma
             )
 
-            if self.accelerated:
-                theta = 1 / np.sqrt(1 + 2*gamma*tau)
-                tau = theta * tau
-                sigma = sigma / theta
-
             self._updatePrimal(
                 out_primal=primal_vars_new,
                 out_fwd=tmp_results_forward,
@@ -2573,6 +2580,11 @@ class PDSoftSenseBaseSolver:
                 tau=tau,
                 theta=theta
                 )
+
+            if self.accelerated:
+                theta = 1 / np.sqrt(1 + 2*gamma*tau)
+                tau = theta * tau
+                sigma = sigma / theta
 
             for j in primal_vars_new:
                 primal_vars[j] = primal_vars_new[j]
@@ -2746,7 +2758,7 @@ class PDSoftSenseBaseSolver:
         """
         return self._prg[idx].update_z_tv(
             self._queue[4*idx+idxq],
-            self._kernelsize, None,  # np.shape(inp[0])[1:-1]
+            self._kernelsize, None,
             outp.data,
             inp[0].data,
             inp[1].data,
@@ -2754,6 +2766,28 @@ class PDSoftSenseBaseSolver:
             np.int32(self.unknowns),
             wait_for=(outp.events + inp[0].events +
                       inp[1].events + wait_for))
+
+    def update_Kyk2(self, outp, inp, par=None, idx=0, idxq=0,
+                    bound_cond=0, wait_for=[]):
+        """
+        Precompute the v-part of the Adjoint Linear operator.
+
+        Args
+        ----
+          div (PyOpenCL.Array): The result of the computation
+          u (PyOpenCL.Array): Dual Variable of the symmetrized gradient of TGV
+          z (PyOpenCL.Array): Dual Variable of the gradient of TGV
+          wait_for (list): A optional list for PyOpenCL.events to wait for
+        """
+        return self._prg[idx].update_Kyk2(
+            self._queue[4*idx+idxq],
+            self._kernelsize, None,
+            outp.data, inp[0].data, inp[1].data,
+            np.int32(self.unknowns),
+            par[idx].data,
+            np.int32(bound_cond),
+            np.float32(self.dz),
+            wait_for=outp.events + inp[0].events + inp[1].events+wait_for)
 
     def update_z1_tgv(self, outp, inp, par=None, idx=0, idxq=0,
                       bound_cond=0, wait_for=[]):
@@ -2767,7 +2801,7 @@ class PDSoftSenseBaseSolver:
           gx (PyOpenCL.Array): Linear gradient operator applied to x
           v (PyOpenCL.Array): Primal variable v
           sigma (float): Dual step size
-          alphainv (float): Inverse regularization parameter of the first
+          alpha (float): Regularization parameter of the first
             TGV functional
[…]
