### 2020-12-11 #2727 @Kevin-Mattheus-Moerman
@MaierOli2010 I have two minor points about the paper:

- [ ] For consistency, can you add "Austria" to the 3rd affiliation?

- [ ] Should the heading "Algorithmic" be "Algorithms" instead?
