### 2020-12-12 #2727 @MaierOli2010
@Kevin-Mattheus-Moerman
Added "Austria" to the 3rd affiliation and renamed "Algorithmic" to "Algorithms"
Both changes are done to "master" branch as I've merged everything there.
