+++ pyqmri/models/FFCPaperPhaseR1.py
@@ -24,31 +24,36 @@ class Model(BaseModel):
             self.t = self.t[None]
 
         self.numT1Scale = len(self.b)
+        self.numC = 1#len(self.b)
+        self.numAlpha = len(self.b)
 
-        par["unknowns_TGV"] = 1 + len(self.b) + len(self.b)
+        par["unknowns_TGV"] = self.numC + self.numAlpha + len(self.b)
         par["unknowns_H1"] = 0
         par["unknowns"] = par["unknowns_TGV"]+par["unknowns_H1"]
 
+
         self.unknowns = par["unknowns"]
         
-        t1min = np.min(self.t)/3
+        t1min = 0 #np.min(self.t)/3
 
         for j in range(par["unknowns"]):
             self.uk_scale.append(1)
-
-        self.constraints.append(
-            constraints(0,
-                        1000,
-                        True))
-        for j in range(self.numT1Scale):
+            
+        for j in range(self.numC):
             self.constraints.append(
                 constraints(0,
-                            1,
+                            1000,
                             False))
+        for j in range(self.numAlpha):
+            self.constraints.append(
+                constraints(0,
+                            2,
+                            True))
         for j in range(self.numT1Scale):
             self.constraints.append(
-                constraints(1/(800/(j+1)),
-                            1/t1min,
+                constraints(
+                            1/(2000/(j+1)),
+                            1,
                             True))
         self._ind1 = 0
         self._ind2 = 0
@@ -59,10 +64,12 @@ class Model(BaseModel):
 
     def rescale(self, x):
         tmp_x = np.copy(x)
-        tmp_x[0] *= self.uk_scale[0]
-        ukname = ["C"]
-        for j in range(self.numT1Scale):
-            tmp_x[1+j] *= self.uk_scale[1+j]
+        ukname = []
+        for j in range(self.numC):
+            tmp_x[j] *= self.uk_scale[j]
+            ukname.append("C_"+str(1+j))
+        for j in range(self.numAlpha):
+            tmp_x[self.numC+j] *= self.uk_scale[self.numC+j]
             ukname.append("alpha_"+str(1+j))
         for j in range(self.numT1Scale):
             tmp_x[-self.numT1Scale+j] *= self.uk_scale[-self.numT1Scale+j]
@@ -84,8 +91,8 @@ class Model(BaseModel):
             offset = len(self.t[j])
             t = self.t[j][:, None, None, None]
             S[offset*(j):offset*(j+1)] = (
-                x[0] * self.uk_scale[0]*x[1+j]*self.uk_scale[1+j]
-                * (-self.b0 *
+                x[np.mod(j,self.numC)] * self.uk_scale[np.mod(j,self.numC)]
+                * (-self.b0 * x[self.numC+np.mod(j,self.numAlpha)]*self.uk_scale[self.numC+np.mod(j,self.numAlpha)] *
                    np.exp(-t * (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j]))
                    + (1 - np.exp(-t * (x[-self.numT1Scale+j] * self.uk_scale[-self.numT1Scale+j])))
                    * self.b[j])
@@ -100,20 +107,20 @@ class Model(BaseModel):
         gradAlpha = self._gradAlpha(x)
         gradT1 = self._gradT1(x)
 
-        grad = np.concatenate((gradC[None], gradAlpha, gradT1), axis=0)
+        grad = np.concatenate((gradC, gradAlpha, gradT1), axis=0)
         return grad
 
     def _gradC(self, x):
[…]
