+++ pyqmri/irgn.py
@@ -366,7 +366,8 @@ class IRGNOptimizer:
 # New .hdf5 save files ########################################################
 ###############################################################################
     def _saveToFile(self, myit, result):
-        f = h5py.File(self.par["outdir"]+"output_" + self.par["fname"], "a")
+        f = h5py.File(self.par["outdir"]+"output_" + self.par["fname"] + ".h5",
+                      "a")
         if self._reg_type == 'TGV':
             f.create_dataset("tgv_result_iter_"+str(myit), result.shape,
                              dtype=self._DTYPE, data=result)
